default to file storage in jetstream
