+++ api/schemas_generated.go
@@ -1,4 +1,4 @@
-// auto generated 2020-11-26 13:47:25.866588 +0100 CET m=+12.868087879
+// auto generated 2020-12-11 10:41:02.085386 +0100 CET m=+12.958694136
 
 package api
 
@@ -948,8 +948,8 @@ func init() {
 	schemas["io.nats.jetstream.advisory.v1.restore_complete"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYWR2aXNvcnkvdjEvcmVzdG9yZV9jb21wbGV0ZS5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQW4gQWR2aXNvcnkgc2VudCB3aGVuIGEgU3RyZWFtIHJlc3RvcmUgaXMgY29tcGxldGVkIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYWR2aXNvcnkudjEucmVzdG9yZV9jb21wbGV0ZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIsCiAgICAiaWQiLAogICAgInRpbWVzdGFtcCIsCiAgICAic3RyZWFtIiwKICAgICJzdGFydCIsCiAgICAiZW5kIiwKICAgICJieXRlcyIsCiAgICAiY2xpZW50IgogIF0sCiAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFkdmlzb3J5LnYxLnJlc3RvcmVfY29tcGxldGUiCiAgICB9LAogICAgImlkIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZGVzY3JpcHRpb24iOiAiVW5pcXVlIGNvcnJlbGF0aW9uIElEIGZvciB0aGlzIGV2ZW50IgogICAgfSwKICAgICJ0aW1lc3RhbXAiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHRoaXMgZXZlbnQgd2FzIGNyZWF0ZWQgaW4gUkZDMzMzOSBmb3JtYXQiCiAgICB9LAogICAgInN0cmVhbSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBTdHJlYW0gYmVpbmcgcmVzdG9yZWQiCiAgICB9LAogICAgInN0YXJ0IjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgdGhlIFJlc3RvcmUgcHJvY2VzcyBzdGFydGVkIiwKICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICB9LAogICAgImVuZCI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHRoZSBSZXN0b3JlIHdhcyBjb21wbGV0ZWQiLAogICAgICAiZm9ybWF0IjogImRhdGUtdGltZSIKICAgIH0sCiAgICAiYnl0ZXMiOiB7CiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG51bWJlciBvZiBieXRlcyB0aGF0IHdhcyByZWNlaXZlZCIsCiAgICAgICJtaW5pbXVtIjogMAogICAgfSwKICAgICJjbGllbnQiOiB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiaG9zdCIsCiAgICAgICAgInBvcnQiLAogICAgICAgICJjaWQiLAogICAgICAgICJhY2NvdW50IgogICAgICBdLAogICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImhvc3QiOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBJUCBhZGRyZXNzIHdoZXJlIHRoZSBjbGllbnQgY29ubmVjdHMgZnJvbSIKICAgICAgICB9LAogICAgICAgICJwb3J0IjogewogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHBvcnQgbnVtYmVyIHdoZXJlIHRoZSBjbGllbnQgY29ubmVjdHMgZnJvbSIKICAgICAgICB9LAogICAgICAgICJjaWQiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdW5pcXVlIGNsaWVudCBJRCB0aGUgc2VydmVyIGFzc2lnbmVkIHRvIHRoZSBjb25uZWN0aW9uIgogICAgICAgIH0sCiAgICAgICAgImFjY291bnQiOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBhY2NvdW50IHRoZSB1c2VyIGJlbG9uZ3MgdG8iCiAgICAgICAgfSwKICAgICAgICAidXNlciI6IHsKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHVzZXJuYW1lIHRoYXQgd2FzIHVzZWQgZHVyaW5nIGF1dGhlbnRpY2F0aW9uLCBpZiBhbnkiCiAgICAgICAgfSwKICAgICAgICAibmFtZSI6IHsKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG5hbWUgdGhlIGNsaWVudCBhc3NpZ25lZCB0byB0aGUgY29ubmVjdGlvbiBkdXJpbmcgY29ubmVjdGlvbiBuZWdvdGlhdGlvbiIKICAgICAgICB9LAogICAgICAgICJsYW5ndWFnZSI6IHsKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGNsaWVudCBsaWJyYXJ5IGxhbmd1YWdlIHVzZWQgdG8gY3JlYXRlIHRoZSBjb25uZWN0aW9uIgogICAgICAgIH0sCiAgICAgICAgInZlcnNpb24iOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB2ZXJzaW9uIGNsaWVudCBsaWJyYXJ5IHVzZWQgdG8gY3JlYXRlIHRoZSBjb25uZWN0aW9uIgogICAgICAgIH0KICAgICAgfQogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.metric.v1.consumer_ack"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vbWV0cmljL3YxL2NvbnN1bWVyX2Fjay5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiTWV0cmljIHB1Ymxpc2hlZCB3aGVuIGEgbWVzc2FnZSB3YXMgYWNrbm93bGVkZ2VkIHRvIGEgY29uc3VtZXIgd2l0aCBBY2sgU2FtcGxpbmcgZW5hYmxlZCIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLm1ldHJpYy52MS5jb25zdW1lcl9hY2siLAogICJ0eXBlIjogIm9iamVjdCIsCiAgInJlcXVpcmVkIjogWwogICAgInR5cGUiLAogICAgImlkIiwKICAgICJ0aW1lc3RhbXAiLAogICAgInN0cmVhbSIsCiAgICAiY29uc3VtZXIiLAogICAgInN0cmVhbV9zZXEiLAogICAgImNvbnN1bWVyX3NlcSIsCiAgICAiYWNrX3RpbWUiLAogICAgImRlbGl2ZXJpZXMiCiAgXSwKICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAicHJvcGVydGllcyI6IHsKICAgICJ0eXBlIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiY29uc3QiOiAiaW8ubmF0cy5qZXRzdHJlYW0ubWV0cmljLnYxLmNvbnN1bWVyX2FjayIKICAgIH0sCiAgICAiaWQiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJVbmlxdWUgY29ycmVsYXRpb24gSUQgZm9yIHRoaXMgZXZlbnQiCiAgICB9LAogICAgInRpbWVzdGFtcCI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiLAogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgdGhpcyBldmVudCB3YXMgY3JlYXRlZCBpbiBSRkMzMzM5IGZvcm1hdCIKICAgIH0sCiAgICAic3RyZWFtIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG5hbWUgb2YgdGhlIHN0cmVhbSB3aGVyZSB0aGUgbWVzc2FnZSBpcyBzdG9yZWQiCiAgICB9LAogICAgImNvbnN1bWVyIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG5hbWUgb2YgdGhlIGNvbnN1bWVyIHdoZXJlIHRoZSBtZXNzYWdlIGlzIGhlbGQiCiAgICB9LAogICAgInN0cmVhbV9zZXEiOiB7CiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAibWluaW11bSI6IDEsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc2VxdWVuY2Ugb2YgdGhlIG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSB0aGF0IHdlcmUgYWNrbm93bGVkZ2VkIgogICAgfSwKICAgICJjb25zdW1lcl9zZXEiOiB7CiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAibWluaW11bSI6IDEsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc2VxdWVuY2Ugb2YgdGhlIG1lc3NhZ2UgaW4gdGhlIGNvbnN1bWVyIHRoYXQgd2VyZSBhY2tub3dsZWRnZWQiCiAgICB9LAogICAgImFja190aW1lIjogewogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgIm1pbmltdW0iOiAxLAogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgaXQgdG9vayBvbiB0aGUgZmluYWwgZGVsaXZlcnkgZm9yIHRoZSBtZXNzYWdlIHRvIGJlIGFja25vd2xlZGdlZCBpbiBuYW5vc2Vjb25kcyIKICAgIH0sCiAgICAiZGVsaXZlcmllcyI6IHsKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogMSwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBudW1iZXIgb2YgZGVsaXZlcmllcyB0aGF0IHdlcmUgYXR0ZW1wdGVkIGJlZm9yZSBiZWluZyBhY2tub3dsZWRnZWQiCiAgICB9CiAgfQp9Cg==")
 	schemas["io.nats.jetstream.api.v1.consumer_configuration"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL2NvbnN1bWVyX2NvbmZpZ3VyYXRpb24uanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIlRoZSBkYXRhIHN0cnVjdHVyZSB0aGF0IGRlc2NyaWJlIHRoZSBjb25maWd1cmF0aW9uIG9mIGEgTkFUUyBKZXRTdHJlYW0gQ29uc3VtZXIiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuY29uc3VtZXJfY29uZmlndXJhdGlvbiIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAiZGVsaXZlcl9wb2xpY3kiLAogICAgImFja19wb2xpY3kiLAogICAgInJlcGxheV9wb2xpY3kiCiAgXSwKICAiYWxsT2YiOiBbCiAgICB7CiAgICAgICJvbmVPZiI6IFsKICAgICAgICB7CiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAiYWxsIgogICAgICAgICAgICAgIF0KICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0sCiAgICAgICAgewogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgImxhc3QiCiAgICAgICAgICAgICAgXQogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfSwKICAgICAgICB7CiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAibmV3IgogICAgICAgICAgICAgIF0KICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0sCiAgICAgICAgewogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiLAogICAgICAgICAgICAib3B0X3N0YXJ0X3NlcSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAiYnlfc3RhcnRfc2VxdWVuY2UiCiAgICAgICAgICAgICAgXQogICAgICAgICAgICB9LAogICAgICAgICAgICAib3B0X3N0YXJ0X3NlcSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0sCiAgICAgICAgewogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiLAogICAgICAgICAgICAib3B0X3N0YXJ0X3RpbWUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgImJ5X3N0YXJ0X3RpbWUiCiAgICAgICAgICAgICAgXQogICAgICAgICAgICB9LAogICAgICAgICAgICAib3B0X3N0YXJ0X3RpbWUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZm9ybWF0IjogImRhdGUtdGltZSIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgXQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAiZHVyYWJsZV9uYW1lIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiQSB1bmlxdWUgbmFtZSBmb3IgYSBkdXJhYmxlIGNvbnN1bWVyIiwKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgInBhdHRlcm4iOiAiXlteLio+XSskIiwKICAgICAgIm1pbkxlbmd0aCI6IDEKICAgIH0sCiAgICAiZGVsaXZlcl9zdWJqZWN0IjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAibWluTGVuZ3RoIjogMQogICAgfSwKICAgICJhY2tfcG9saWN5IjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZW51bSI6IFsKICAgICAgICAibm9uZSIsCiAgICAgICAgImFsbCIsCiAgICAgICAgImV4cGxpY2l0IgogICAgICBdCiAgICB9LAogICAgImFja193YWl0IjogewogICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGxvbmcgKGluIG5hbm9zZWNvbmRzKSB0byBhbGxvdyBtZXNzYWdlcyB0byByZW1haW4gdW4tYWNrbm93bGVkZ2VkIGJlZm9yZSBhdHRlbXB0aW5nIHJlZGVsaXZlcnkiLAogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgIm1pbmltdW0iOiAxCiAgICB9LAogICAgIm1heF9kZWxpdmVyIjogewogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgIm1pbmltdW0iOiAtMQogICAgfSwKICAgICJmaWx0ZXJfc3ViamVjdCI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgfSwKICAgICJyZXBsYXlfcG9saWN5IjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZW51bSI6IFsKICAgICAgICAiaW5zdGFudCIsCiAgICAgICAgIm9yaWdpbmFsIgogICAgICBdCiAgICB9LAogICAgInNhbXBsZV9mcmVxIjogewogICAgICAidHlwZSI6ICJzdHJpbmciCiAgICB9LAogICAgInJhdGVfbGltaXRfYnBzIjogewogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSByYXRlIGF0IHdoaWNoIG1lc3NhZ2VzIHdpbGwgYmUgZGVsaXZlcmVkIHRvIGNsaWVudHMsIGV4cHJlc3NlZCBpbiBiaXQgcGVyIHNlY29uZCIsCiAgICAgICJtaW5pbXVtIjogMAogICAgfSwKICAgICJtYXhfYWNrX3BlbmRpbmciOiB7CiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG1heGltdW0gbnVtYmVyIG9mIG1lc3NhZ2VzIHdpdGhvdXQgYWNrbm93bGVkZ2VtZW50IHRoYXQgY2FuIGJlIG91dHN0YW5kaW5nLCBvbmNlIHRoaXMgbGltaXQgaXMgcmVhY2hlZCBtZXNzYWdlIGRlbGl2ZXJ5IHdpbGwgYmUgc3VzcGVuZGVkIiwKICAgICAgIm1pbmltdW0iOiAwCiAgICB9CiAgfQp9Cg==")
-	schemas["io.nats.jetstream.api.v1.stream_configuration"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9jb25maWd1cmF0aW9uLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJUaGUgZGF0YSBzdHJ1Y3R1cmUgdGhhdCBkZXNjcmliZSB0aGUgY29uZmlndXJhdGlvbiBvZiBhIE5BVFMgSmV0U3RyZWFtIFN0cmVhbSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fY29uZmlndXJhdGlvbiIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAic3ViamVjdHMiLAogICAgInJldGVudGlvbiIsCiAgICAibWF4X2NvbnN1bWVycyIsCiAgICAibWF4X21zZ3MiLAogICAgIm1heF9ieXRlcyIsCiAgICAibWF4X2FnZSIsCiAgICAic3RvcmFnZSIsCiAgICAibnVtX3JlcGxpY2FzIgogIF0sCiAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgInByb3BlcnRpZXMiOiB7CiAgICAibmFtZSI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICJtaW5MZW5ndGgiOiAwCiAgICB9LAogICAgInN1YmplY3RzIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAiaXRlbXMiOiB7CiAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICB9CiAgICB9LAogICAgInJldGVudGlvbiI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZW51bSI6IFsKICAgICAgICAibGltaXRzIiwKICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICJ3b3JrcXVldWUiCiAgICAgIF0sCiAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgIH0sCiAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICJkZWZhdWx0IjogLTEKICAgIH0sCiAgICAibWF4X21zZ3MiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgImRlZmF1bHQiOiAtMQogICAgfSwKICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAibWluaW11bSI6IC0xLAogICAgICAiZGVmYXVsdCI6IC0xCiAgICB9LAogICAgIm1heF9hZ2UiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogMCwKICAgICAgImRlZmF1bHQiOiAwCiAgICB9LAogICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICJkZWZhdWx0IjogLTEKICAgIH0sCiAgICAic3RvcmFnZSI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJlbnVtIjogWwogICAgICAgICJmaWxlIiwKICAgICAgICAibWVtb3J5IgogICAgICBdLAogICAgICAiZGVmYXVsdCI6ICJtZW1vcnkiCiAgICB9LAogICAgIm51bV9yZXBsaWNhcyI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IHJlcGxpY2FzIHRvIGtlZXAgZm9yIGVhY2ggbWVzc2FnZS4iLAogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgIm1pbmltdW0iOiAxLAogICAgICAiZGVmYXVsdCI6IDEKICAgIH0sCiAgICAibm9fYWNrIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiRGlzYWJsZXMgYWNrbm93bGVkZ2luZyBtZXNzYWdlcyB0aGF0IGFyZSByZWNlaXZlZCBieSB0aGUgU3RyZWFtLiIsCiAgICAgICJ0eXBlIjogImJvb2xlYW4iLAogICAgICAiZGVmYXVsdCI6IGZhbHNlCiAgICB9LAogICAgInRlbXBsYXRlX293bmVyIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiB0aGUgU3RyZWFtIGlzIG1hbmFnZWQgYnkgYSBTdHJlYW0gVGVtcGxhdGUgdGhpcyBpZGVudGlmaWVzIHRoZSB0ZW1wbGF0ZSB0aGF0IG1hbmFnZXMgdGhlIFN0cmVhbS4iLAogICAgICAidHlwZSI6ICJzdHJpbmciCiAgICB9LAogICAgImRpc2NhcmQiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIGEgU3RyZWFtIHJlYWNoIGl0J3MgbGltaXRzIGVpdGhlciBvbGQgbWVzc2FnZXMgYXJlIGRlbGV0ZWQgb3IgbmV3IG9uZXMgYXJlIGRlbmllZCIsCiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJlbnVtIjogWwogICAgICAgICJvbGQiLAogICAgICAgICJuZXciCiAgICAgIF0sCiAgICAgICJkZWZhdWx0IjogIm9sZCIKICAgIH0sCiAgICAiZHVwbGljYXRlX3dpbmRvdyI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHdpbmRvdyB0byB0cmFjayBkdXBsaWNhdGUgbWVzc2FnZXMgZm9yLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciBkZWZhdWx0IiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogMCwKICAgICAgImRlZmF1bHQiOiAwCiAgICB9CiAgfQp9Cg==")
-	schemas["io.nats.jetstream.api.v1.stream_template_configuration"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9jb25maWd1cmF0aW9uLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJUaGUgZGF0YSBzdHJ1Y3R1cmUgdGhhdCBkZXNjcmliZSB0aGUgY29uZmlndXJhdGlvbiBvZiBhIE5BVFMgSmV0U3RyZWFtIFN0cmVhbSBUZW1wbGF0ZSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfY29uZmlndXJhdGlvbiIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAibmFtZSIsCiAgICAiY29uZmlnIiwKICAgICJtYXhfc3RyZWFtcyIKICBdLAogICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICJwcm9wZXJ0aWVzIjogewogICAgIm5hbWUiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtIFRlbXBsYXRlLiIsCiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0rJCIsCiAgICAgICJtaW5MZW5ndGgiOiAxCiAgICB9LAogICAgIm1heF9zdHJlYW1zIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG1heGltdW0gbnVtYmVyIG9mIFN0cmVhbXMgdGhpcyBUZW1wbGF0ZSBjYW4gY3JlYXRlLCAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAibWluaW11bSI6IC0xLAogICAgICAiZGVmYXVsdCI6IC0xCiAgICB9LAogICAgImNvbmZpZyI6IHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJzdWJqZWN0cyIsCiAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICJtYXhfbXNncyIsCiAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICJzdG9yYWdlIiwKICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICBdLAogICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSB1bmlxdWUgbmFtZSBmb3IgdGhlIFN0cmVhbSwgZW1wdHkgZm9yIFN0cmVhbSBUZW1wbGF0ZXMuIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgIm1pbkxlbmd0aCI6IDAKICAgICAgICB9LAogICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGxpc3Qgb2Ygc3ViamVjdHMgdG8gY29uc3VtZSwgc3VwcG9ydHMgd2lsZGNhcmRzLiIsCiAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICJpdGVtcyI6IHsKICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgIH0KICAgICAgICB9LAogICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1lc3NhZ2VzIGFyZSByZXRhaW5lZCBpbiB0aGUgU3RyZWFtLCBvbmNlIHRoaXMgaXMgZXhjZWVkZWQgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICJsaW1pdHMiLAogICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgXSwKICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICB9LAogICAgICAgICJtYXhfY29uc3VtZXJzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICB9LAogICAgICAgICJtYXhfbXNncyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgIH0sCiAgICAgICAgIm1heF9ieXRlcyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgfSwKICAgICAgICAibWF4X2FnZSI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgfSwKICAgICAgICAibWF4X21zZ19zaXplIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICB9LAogICAgICAgICJzdG9yYWdlIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgIF0sCiAgICAgICAgICAiZGVmYXVsdCI6ICJtZW1vcnkiCiAgICAgICAgfSwKICAgICAgICAibnVtX3JlcGxpY2FzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IHJlcGxpY2FzIHRvIGtlZXAgZm9yIGVhY2ggbWVzc2FnZS4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDEsCiAgICAgICAgICAiZGVmYXVsdCI6IDEKICAgICAgICB9LAogICAgICAgICJub19hY2siOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGlzYWJsZXMgYWNrbm93bGVkZ2luZyBtZXNzYWdlcyB0aGF0IGFyZSByZWNlaXZlZCBieSB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgICAgICJkZWZhdWx0IjogZmFsc2UKICAgICAgICB9LAogICAgICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIHRoZSBTdHJlYW0gaXMgbWFuYWdlZCBieSBhIFN0cmVhbSBUZW1wbGF0ZSB0aGlzIGlkZW50aWZpZXMgdGhlIHRlbXBsYXRlIHRoYXQgbWFuYWdlcyB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgfSwKICAgICAgICAiZGlzY2FyZCI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIGEgU3RyZWFtIHJlYWNoIGl0J3MgbGltaXRzIGVpdGhlciBvbGQgbWVzc2FnZXMgYXJlIGRlbGV0ZWQgb3IgbmV3IG9uZXMgYXJlIGRlbmllZCIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICJvbGQiLAogICAgICAgICAgICAibmV3IgogICAgICAgICAgXSwKICAgICAgICAgICJkZWZhdWx0IjogIm9sZCIKICAgICAgICB9LAogICAgICAgICJkdXBsaWNhdGVfd2luZG93IjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHdpbmRvdyB0byB0cmFjayBkdXBsaWNhdGUgbWVzc2FnZXMgZm9yLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciBkZWZhdWx0IiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgfQogICAgICB9CiAgICB9CiAgfQp9Cg==")
+	schemas["io.nats.jetstream.api.v1.stream_configuration"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9jb25maWd1cmF0aW9uLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJUaGUgZGF0YSBzdHJ1Y3R1cmUgdGhhdCBkZXNjcmliZSB0aGUgY29uZmlndXJhdGlvbiBvZiBhIE5BVFMgSmV0U3RyZWFtIFN0cmVhbSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fY29uZmlndXJhdGlvbiIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAic3ViamVjdHMiLAogICAgInJldGVudGlvbiIsCiAgICAibWF4X2NvbnN1bWVycyIsCiAgICAibWF4X21zZ3MiLAogICAgIm1heF9ieXRlcyIsCiAgICAibWF4X2FnZSIsCiAgICAic3RvcmFnZSIsCiAgICAibnVtX3JlcGxpY2FzIgogIF0sCiAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgInByb3BlcnRpZXMiOiB7CiAgICAibmFtZSI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICJtaW5MZW5ndGgiOiAwCiAgICB9LAogICAgInN1YmplY3RzIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAiaXRlbXMiOiB7CiAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICB9CiAgICB9LAogICAgInJldGVudGlvbiI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZW51bSI6IFsKICAgICAgICAibGltaXRzIiwKICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICJ3b3JrcXVldWUiCiAgICAgIF0sCiAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgIH0sCiAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICJkZWZhdWx0IjogLTEKICAgIH0sCiAgICAibWF4X21zZ3MiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgImRlZmF1bHQiOiAtMQogICAgfSwKICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAibWluaW11bSI6IC0xLAogICAgICAiZGVmYXVsdCI6IC0xCiAgICB9LAogICAgIm1heF9hZ2UiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogMCwKICAgICAgImRlZmF1bHQiOiAwCiAgICB9LAogICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICJkZWZhdWx0IjogLTEKICAgIH0sCiAgICAic3RvcmFnZSI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJlbnVtIjogWwogICAgICAgICJmaWxlIiwKICAgICAgICAibWVtb3J5IgogICAgICBdLAogICAgICAiZGVmYXVsdCI6ICJmaWxlIgogICAgfSwKICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJtaW5pbXVtIjogMSwKICAgICAgImRlZmF1bHQiOiAxCiAgICB9LAogICAgIm5vX2FjayI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgfSwKICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgfSwKICAgICJkaXNjYXJkIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiZW51bSI6IFsKICAgICAgICAib2xkIiwKICAgICAgICAibmV3IgogICAgICBdLAogICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICB9LAogICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAibWluaW11bSI6IDAsCiAgICAgICJkZWZhdWx0IjogMAogICAgfQogIH0KfQo=")
+	schemas["io.nats.jetstream.api.v1.stream_template_configuration"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9jb25maWd1cmF0aW9uLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJUaGUgZGF0YSBzdHJ1Y3R1cmUgdGhhdCBkZXNjcmliZSB0aGUgY29uZmlndXJhdGlvbiBvZiBhIE5BVFMgSmV0U3RyZWFtIFN0cmVhbSBUZW1wbGF0ZSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfY29uZmlndXJhdGlvbiIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAibmFtZSIsCiAgICAiY29uZmlnIiwKICAgICJtYXhfc3RyZWFtcyIKICBdLAogICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICJwcm9wZXJ0aWVzIjogewogICAgIm5hbWUiOiB7CiAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtIFRlbXBsYXRlLiIsCiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0rJCIsCiAgICAgICJtaW5MZW5ndGgiOiAxCiAgICB9LAogICAgIm1heF9zdHJlYW1zIjogewogICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG1heGltdW0gbnVtYmVyIG9mIFN0cmVhbXMgdGhpcyBUZW1wbGF0ZSBjYW4gY3JlYXRlLCAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAibWluaW11bSI6IC0xLAogICAgICAiZGVmYXVsdCI6IC0xCiAgICB9LAogICAgImNvbmZpZyI6IHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJzdWJqZWN0cyIsCiAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICJtYXhfbXNncyIsCiAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICJzdG9yYWdlIiwKICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICBdLAogICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSB1bmlxdWUgbmFtZSBmb3IgdGhlIFN0cmVhbSwgZW1wdHkgZm9yIFN0cmVhbSBUZW1wbGF0ZXMuIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgIm1pbkxlbmd0aCI6IDAKICAgICAgICB9LAogICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGxpc3Qgb2Ygc3ViamVjdHMgdG8gY29uc3VtZSwgc3VwcG9ydHMgd2lsZGNhcmRzLiIsCiAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICJpdGVtcyI6IHsKICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgIH0KICAgICAgICB9LAogICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1lc3NhZ2VzIGFyZSByZXRhaW5lZCBpbiB0aGUgU3RyZWFtLCBvbmNlIHRoaXMgaXMgZXhjZWVkZWQgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICJsaW1pdHMiLAogICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgXSwKICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICB9LAogICAgICAgICJtYXhfY29uc3VtZXJzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICB9LAogICAgICAgICJtYXhfbXNncyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgIH0sCiAgICAgICAgIm1heF9ieXRlcyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgfSwKICAgICAgICAibWF4X2FnZSI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgfSwKICAgICAgICAibWF4X21zZ19zaXplIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICB9LAogICAgICAgICJzdG9yYWdlIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgIF0sCiAgICAgICAgICAiZGVmYXVsdCI6ICJmaWxlIgogICAgICAgIH0sCiAgICAgICAgIm51bV9yZXBsaWNhcyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAxLAogICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgfSwKICAgICAgICAibm9fYWNrIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICAgICAiZGVmYXVsdCI6IGZhbHNlCiAgICAgICAgfSwKICAgICAgICAidGVtcGxhdGVfb3duZXIiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiB0aGUgU3RyZWFtIGlzIG1hbmFnZWQgYnkgYSBTdHJlYW0gVGVtcGxhdGUgdGhpcyBpZGVudGlmaWVzIHRoZSB0ZW1wbGF0ZSB0aGF0IG1hbmFnZXMgdGhlIFN0cmVhbS4iLAogICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgIH0sCiAgICAgICAgImRpc2NhcmQiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgIm5ldyIKICAgICAgICAgIF0sCiAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgfSwKICAgICAgICAiZHVwbGljYXRlX3dpbmRvdyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.api.v1.account_info_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL2FjY291bnRfaW5mb19yZXNwb25zZS5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXNwb25zZSBmcm9tIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5JTkZPIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5hY2NvdW50X2luZm9fcmVzcG9uc2UiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgInJlcXVpcmVkIjogWwogICAgInR5cGUiCiAgXSwKICAib25lT2YiOiBbCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAibWVtb3J5IiwKICAgICAgICAic3RvcmFnZSIsCiAgICAgICAgInN0cmVhbXMiLAogICAgICAgICJsaW1pdHMiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJtZW1vcnkiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWVtb3J5IFN0b3JhZ2UgYmVpbmcgdXNlZCBmb3IgU3RyZWFtIE1lc3NhZ2Ugc3RvcmFnZSIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0sCiAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRmlsZSBTdG9yYWdlIGJlaW5nIHVzZWQgZm9yIFN0cmVhbSBNZXNzYWdlIHN0b3JhZ2UiLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICB9LAogICAgICAgICJzdHJlYW1zIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBhY3RpdmUgU3RyZWFtcyIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0sCiAgICAgICAgImxpbWl0cyI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICAgICAibWF4X21lbW9yeSIsCiAgICAgICAgICAgICJtYXhfc3RvcmFnZSIsCiAgICAgICAgICAgICJtYXhfc3RyZWFtcyIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgIm1heF9tZW1vcnkiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBtYXhpbXVtIGFtb3VudCBvZiBNZW1vcnkgc3RvcmFnZSBTdHJlYW0gTWVzc2FnZXMgbWF5IGNvbnN1bWUiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogLTEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm1heF9zdG9yYWdlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbWF4aW11bSBhbW91bnQgb2YgRmlsZSBzdG9yYWdlIFN0cmVhbSBNZXNzYWdlcyBtYXkgY29uc3VtZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMQogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X3N0cmVhbXMiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBtYXhpbXVtIG51bWJlciBvZiBTdHJlYW1zIGFuIGFjY291bnQgY2FuIGNyZWF0ZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMQogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG1heGltdW0gbnVtYmVyIG9mIENvbnN1bWVyIGFuIGFjY291bnQgY2FuIGNyZWF0ZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMQogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgImVycm9yIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiZXJyb3IiOiB7CiAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29kZSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImNvZGUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDMwMCwKICAgICAgICAgICAgICAibWF4aW11bSI6IDY5OQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGVzY3JpcHRpb24iOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuYWNjb3VudF9pbmZvX3Jlc3BvbnNlIgogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.api.v1.consumer_create_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL2NvbnN1bWVyX2NyZWF0ZV9yZXF1ZXN0Lmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlcXVlc3QgdG8gdGhlIEpldFN0cmVhbSAkSlMuQVBJLkNPTlNVTUVSLkNSRUFURSBhbmQgJEpTLkFQSS5DT05TVU1FUi5EVVJBQkxFLkNSRUFURSBBUElzIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLmNvbnN1bWVyX2NyZWF0ZV9yZXF1ZXN0IiwKICAidHlwZSI6ICJvYmplY3QiLAogICJyZXF1aXJlZCI6IFsKICAgICJzdHJlYW1fbmFtZSIsCiAgICAiY29uZmlnIgogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAic3RyZWFtX25hbWUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbmFtZSBvZiB0aGUgc3RyZWFtIHRvIGNyZWF0ZSB0aGUgY29uc3VtZXIgaW4iCiAgICB9LAogICAgImNvbmZpZyI6IHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBjb25zdW1lciBjb25maWd1cmF0aW9uIiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJkZWxpdmVyX3BvbGljeSIsCiAgICAgICAgImFja19wb2xpY3kiLAogICAgICAgICJyZXBsYXlfcG9saWN5IgogICAgICBdLAogICAgICAiYWxsT2YiOiBbCiAgICAgICAgewogICAgICAgICAgIm9uZU9mIjogWwogICAgICAgICAgICB7CiAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICJhbGwiCiAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgIHsKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiCiAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImxhc3QiCiAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgIHsKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiCiAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgIm5ldyIKICAgICAgICAgICAgICAgICAgXQogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfSwKICAgICAgICAgICAgewogICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSIsCiAgICAgICAgICAgICAgICAib3B0X3N0YXJ0X3NlcSIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAiYnlfc3RhcnRfc2VxdWVuY2UiCiAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAib3B0X3N0YXJ0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfSwKICAgICAgICAgICAgewogICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSIsCiAgICAgICAgICAgICAgICAib3B0X3N0YXJ0X3RpbWUiCiAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImJ5X3N0YXJ0X3RpbWUiCiAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAib3B0X3N0YXJ0X3RpbWUiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfQogICAgICAgICAgXQogICAgICAgIH0KICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImR1cmFibGVfbmFtZSI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciBhIGR1cmFibGUgY29uc3VtZXIiLAogICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0rJCIsCiAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgIH0sCiAgICAgICAgImRlbGl2ZXJfc3ViamVjdCI6IHsKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgIH0sCiAgICAgICAgImFja19wb2xpY3kiOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICJub25lIiwKICAgICAgICAgICAgImFsbCIsCiAgICAgICAgICAgICJleHBsaWNpdCIKICAgICAgICAgIF0KICAgICAgICB9LAogICAgICAgICJhY2tfd2FpdCI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbG9uZyAoaW4gbmFub3NlY29uZHMpIHRvIGFsbG93IG1lc3NhZ2VzIHRvIHJlbWFpbiB1bi1hY2tub3dsZWRnZWQgYmVmb3JlIGF0dGVtcHRpbmcgcmVkZWxpdmVyeSIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMQogICAgICAgIH0sCiAgICAgICAgIm1heF9kZWxpdmVyIjogewogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IC0xCiAgICAgICAgfSwKICAgICAgICAiZmlsdGVyX3N1YmplY3QiOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgfSwKICAgICAgICAicmVwbGF5X3BvbGljeSI6IHsKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgImluc3RhbnQiLAogICAgICAgICAgICAib3JpZ2luYWwiCiAgICAgICAgICBdCiAgICAgICAgfSwKICAgICAgICAic2FtcGxlX2ZyZXEiOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgfSwKICAgICAgICAicmF0ZV9saW1pdF9icHMiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgcmF0ZSBhdCB3aGljaCBtZXNzYWdlcyB3aWxsIGJlIGRlbGl2ZXJlZCB0byBjbGllbnRzLCBleHByZXNzZWQgaW4gYml0IHBlciBzZWNvbmQiLAogICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgfSwKICAgICAgICAibWF4X2Fja19wZW5kaW5nIjogewogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG1heGltdW0gbnVtYmVyIG9mIG1lc3NhZ2VzIHdpdGhvdXQgYWNrbm93bGVkZ2VtZW50IHRoYXQgY2FuIGJlIG91dHN0YW5kaW5nLCBvbmNlIHRoaXMgbGltaXQgaXMgcmVhY2hlZCBtZXNzYWdlIGRlbGl2ZXJ5IHdpbGwgYmUgc3VzcGVuZGVkIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.api.v1.consumer_create_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL2NvbnN1bWVyX2NyZWF0ZV9yZXNwb25zZS5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXNwb25zZSBmcm9tIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5DT05TVU1FUi5DUkVBVEUgQVBJIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLmNvbnN1bWVyX2NyZWF0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJzdHJlYW1fbmFtZSIsCiAgICAgICAgIm5hbWUiLAogICAgICAgICJjb25maWciLAogICAgICAgICJjcmVhdGVkIiwKICAgICAgICAiZGVsaXZlcmVkIiwKICAgICAgICAiYWNrX2Zsb29yIiwKICAgICAgICAibnVtX2Fja19wZW5kaW5nIiwKICAgICAgICAibnVtX3JlZGVsaXZlcmVkIiwKICAgICAgICAibnVtX3dhaXRpbmciLAogICAgICAgICJudW1fcGVuZGluZyIKICAgICAgXSwKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgInN0cmVhbV9uYW1lIjogewogICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgU3RyZWFtIHRoZSBjb25zdW1lciBiZWxvbmdzIHRvIgogICAgICAgIH0sCiAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBjb25zdW1lciwgZWl0aGVyIG1hY2hpbmUgZ2VuZXJhdGVkIG9yIHRoZSBkdXJhYmxlIG5hbWUiCiAgICAgICAgfSwKICAgICAgICAiY29uZmlnIjogewogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiLAogICAgICAgICAgICAiYWNrX3BvbGljeSIsCiAgICAgICAgICAgICJyZXBsYXlfcG9saWN5IgogICAgICAgICAgXSwKICAgICAgICAgICJhbGxPZiI6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICJvbmVPZiI6IFsKICAgICAgICAgICAgICAgIHsKICAgICAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IjogewogICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAiYWxsIgogICAgICAgICAgICAgICAgICAgICAgXQogICAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgIHsKICAgICAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IjogewogICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAibGFzdCIKICAgICAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiCiAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgICAgICJkZWxpdmVyX3BvbGljeSI6IHsKICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAgICAgIm5ldyIKICAgICAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiLAogICAgICAgICAgICAgICAgICAgICJvcHRfc3RhcnRfc2VxIgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiOiB7CiAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgICAgICJieV9zdGFydF9zZXF1ZW5jZSIKICAgICAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICJvcHRfc3RhcnRfc2VxIjogewogICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICAgICAiZGVsaXZlcl9wb2xpY3kiLAogICAgICAgICAgICAgICAgICAgICJvcHRfc3RhcnRfdGltZSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgImRlbGl2ZXJfcG9saWN5IjogewogICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAiYnlfc3RhcnRfdGltZSIKICAgICAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICJvcHRfc3RhcnRfdGltZSI6IHsKICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgICAiZm9ybWF0IjogImRhdGUtdGltZSIKICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICBdCiAgICAgICAgICAgIH0KICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImR1cmFibGVfbmFtZSI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSB1bmlxdWUgbmFtZSBmb3IgYSBkdXJhYmxlIGNvbnN1bWVyIiwKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0rJCIsCiAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlbGl2ZXJfc3ViamVjdCI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJhY2tfcG9saWN5IjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAibm9uZSIsCiAgICAgICAgICAgICAgICAiYWxsIiwKICAgICAgICAgICAgICAgICJleHBsaWNpdCIKICAgICAgICAgICAgICBdCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJhY2tfd2FpdCI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGxvbmcgKGluIG5hbm9zZWNvbmRzKSB0byBhbGxvdyBtZXNzYWdlcyB0byByZW1haW4gdW4tYWNrbm93bGVkZ2VkIGJlZm9yZSBhdHRlbXB0aW5nIHJlZGVsaXZlcnkiLAogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMQogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X2RlbGl2ZXIiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZmlsdGVyX3N1YmplY3QiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICB9LAogICAgICAgICAgICAicmVwbGF5X3BvbGljeSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgImluc3RhbnQiLAogICAgICAgICAgICAgICAgIm9yaWdpbmFsIgogICAgICAgICAgICAgIF0KICAgICAgICAgICAgfSwKICAgICAgICAgICAgInNhbXBsZV9mcmVxIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgInJhdGVfbGltaXRfYnBzIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgcmF0ZSBhdCB3aGljaCBtZXNzYWdlcyB3aWxsIGJlIGRlbGl2ZXJlZCB0byBjbGllbnRzLCBleHByZXNzZWQgaW4gYml0IHBlciBzZWNvbmQiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X2Fja19wZW5kaW5nIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbWF4aW11bSBudW1iZXIgb2YgbWVzc2FnZXMgd2l0aG91dCBhY2tub3dsZWRnZW1lbnQgdGhhdCBjYW4gYmUgb3V0c3RhbmRpbmcsIG9uY2UgdGhpcyBsaW1pdCBpcyByZWFjaGVkIG1lc3NhZ2UgZGVsaXZlcnkgd2lsbCBiZSBzdXNwZW5kZWQiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfSwKICAgICAgICAiY3JlYXRlZCI6IHsKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZm9ybWF0IjogImRhdGUtdGltZSIKICAgICAgICB9LAogICAgICAgICJkZWxpdmVyZWQiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGxhc3QgbWVzc2FnZSBkZWxpdmVyZWQgZnJvbSB0aGlzIENvbnN1bWVyIiwKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvbnN1bWVyX3NlcSIsCiAgICAgICAgICAgICJzdHJlYW1fc2VxIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29uc3VtZXJfc2VxIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc2VxdWVuY2UgbnVtYmVyIG9mIHRoZSBDb25zdW1lciIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJzdHJlYW1fc2VxIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc2VxdWVuY2UgbnVtYmVyIG9mIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfSwKICAgICAgICAiYWNrX2Zsb29yIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBoaWdoZXN0IGNvbnRpZ3VvdXMgYWNrbm93bGVkZ2VkIG1lc3NhZ2UiLAogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29uc3VtZXJfc2VxIiwKICAgICAgICAgICAgInN0cmVhbV9zZXEiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb25zdW1lcl9zZXEiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzZXF1ZW5jZSBudW1iZXIgb2YgdGhlIENvbnN1bWVyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgfSwKICAgICAgICAgICAgInN0cmVhbV9zZXEiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzZXF1ZW5jZSBudW1iZXIgb2YgdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9LAogICAgICAgICJudW1fYWNrX3BlbmRpbmciOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG51bWJlciBvZiBtZXNzYWdlcyBwZW5kaW5nIGFja25vd2xlZGdlbWVudCIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0sCiAgICAgICAgIm51bV9yZWRlbGl2ZXJlZCI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbnVtYmVyIG9mIHJlZGVsaXZlcmllcyB0aGF0IGhhdmUgYmVlbiBwZXJmb3JtZWQiLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICB9LAogICAgICAgICJudW1fd2FpdGluZyI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbnVtYmVyIG9mIHB1bGwgY29uc3VtZXJzIHdhaXRpbmcgZm9yIG1lc3NhZ2VzIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgfSwKICAgICAgICAibnVtX3BlbmRpbmciOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIG51bWJlciBvZiBtZXNzYWdlcyBsZWZ0IHVuY29uc3VtZWQgaW4gdGhpcyBDb25zdW1lciIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfSwKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9CiAgXSwKICAicHJvcGVydGllcyI6IHsKICAgICJ0eXBlIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiY29uc3QiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLmNvbnN1bWVyX2NyZWF0ZV9yZXNwb25zZSIKICAgIH0KICB9Cn0K")
@@ -960,12 +960,12 @@ func init() {
 	schemas["io.nats.jetstream.api.v1.consumer_names_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL2NvbnN1bWVyX25hbWVzX3JlcXVlc3QuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVxdWVzdCB0byB0aGUgSmV0U3RyZWFtICRKUy5BUEkuQ09OU1VNRVIuTkFNRVMgQVBJIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLmNvbnN1bWVyX25hbWVzX3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgImFsbE9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJvZmZzZXQiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJvZmZzZXQiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAic3ViamVjdCI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImRlc2NyaXB0aW9uIjogIkZpbHRlciB0aGUgbmFtZXMgdG8gdGhvc2UgY29uc3VtaW5nIG1lc3NhZ2VzIG1hdGNoaW5nIHRoaXMgc3ViamVjdCBvciB3aWxkY2FyZCIKICAgIH0KICB9Cn0K")
 	schemas["io.nats.jetstream.api.v1.consumer_names_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL2NvbnN1bWVyX25hbWVzX3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLkNPTlNVTUVSLk5BTUVTIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5jb25zdW1lcl9uYW1lc19yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJhbGxPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJ0b3RhbCIsCiAgICAgICAgIm9mZnNldCIsCiAgICAgICAgImxpbWl0IgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAidG90YWwiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0sCiAgICAgICAgIm9mZnNldCI6IHsKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgfSwKICAgICAgICAibGltaXQiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgIm9uZU9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgImVycm9yIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiZXJyb3IiOiB7CiAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29kZSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImNvZGUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDMwMCwKICAgICAgICAgICAgICAibWF4aW11bSI6IDY5OQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGVzY3JpcHRpb24iOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiY29uc3VtZXJzIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiY29uc3VtZXJzIjogewogICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9CiAgXSwKICAicHJvcGVydGllcyI6IHsKICAgICJ0eXBlIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiY29uc3QiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLmNvbnN1bWVyX25hbWVzX3Jlc3BvbnNlIgogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.api.v1.consumer_getnext_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL2NvbnN1bWVyX2dldG5leHRfcmVxdWVzdC5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXF1ZXN0IHRvIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5DT05TVU1FUi5NU0cuTkVYVCBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuY29uc3VtZXJfZ2V0bmV4dF9yZXF1ZXN0IiwKICAidHlwZSI6ICJvYmplY3QiLAogICJwcm9wZXJ0aWVzIjogewogICAgImV4cGlyZXMiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHRpbWUgd2hlbiB0aGlzIHJlcXVlc3Qgc2hvdWxkIGJlIGV4cGlyZWQgZnJvbSB0aGUgc2VydmVyIHdhaXQgbGlzdCIsCiAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgfSwKICAgICJiYXRjaCI6IHsKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyB0aGUgc2VydmVyIHNob3VsZCBkZWxpdmVyIHRvIHRoZSByZXF1ZXN0b3IiLAogICAgICAibWluaW11bSI6IDAKICAgIH0sCiAgICAibm9fd2FpdCI6IHsKICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIHRydWUgYSByZXNwb25zZSB3aXRoIGEgNDA0IHN0YXR1cyBoZWFkZXIgd2lsbCBiZSByZXR1cm5lZCB3aGVuIG5vIG1lc3NhZ2VzIGFyZSBhdmFpbGFibGUiCiAgICB9CiAgfQp9Cg==")
-	schemas["io.nats.jetstream.api.v1.stream_create_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9jcmVhdGVfcmVxdWVzdC5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXF1ZXN0IHRvIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5TVFJFQU0uQ1JFQVRFIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fY3JlYXRlX3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgImFsbE9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAicmV0ZW50aW9uIiwKICAgICAgICAibWF4X2NvbnN1bWVycyIsCiAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAibWF4X2J5dGVzIiwKICAgICAgICAibWF4X2FnZSIsCiAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICJudW1fcmVwbGljYXMiCiAgICAgIF0sCiAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAibmFtZSI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgIH0sCiAgICAgICAgInN1YmplY3RzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICJ0eXBlIjogImFycmF5IiwKICAgICAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgICAgfQogICAgICAgIH0sCiAgICAgICAgInJldGVudGlvbiI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICJpbnRlcmVzdCIsCiAgICAgICAgICAgICJ3b3JrcXVldWUiCiAgICAgICAgICBdLAogICAgICAgICAgImRlZmF1bHQiOiAibGltaXRzIgogICAgICAgIH0sCiAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgQ29uc3VtZXJzIGNhbiBiZSBkZWZpbmVkIGZvciBhIGdpdmVuIFN0cmVhbS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgIH0sCiAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IG1lc3NhZ2VzIG1heSBiZSBpbiBhIFN0cmVhbSwgb2xkZXN0IG1lc3NhZ2VzIHdpbGwgYmUgcmVtb3ZlZCBpZiB0aGUgU3RyZWFtIGV4Y2VlZHMgdGhpcyBzaXplLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgfSwKICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBiaWcgdGhlIFN0cmVhbSBtYXkgYmUsIHdoZW4gdGhlIGNvbWJpbmVkIHN0cmVhbSBzaXplIGV4Y2VlZHMgdGhpcyBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICB9LAogICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIk1heGltdW0gYWdlIG9mIGFueSBtZXNzYWdlIGluIHRoZSBzdHJlYW0sIGV4cHJlc3NlZCBpbiBtaWNyb3NlY29uZHMuIDAgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAiZGVmYXVsdCI6IDAKICAgICAgICB9LAogICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGxhcmdlc3QgbWVzc2FnZSB0aGF0IHdpbGwgYmUgYWNjZXB0ZWQgYnkgdGhlIFN0cmVhbS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgIH0sCiAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHN0b3JhZ2UgYmFja2VuZCB0byB1c2UgZm9yIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgImZpbGUiLAogICAgICAgICAgICAibWVtb3J5IgogICAgICAgICAgXSwKICAgICAgICAgICJkZWZhdWx0IjogIm1lbW9yeSIKICAgICAgICB9LAogICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgcmVwbGljYXMgdG8ga2VlcCBmb3IgZWFjaCBtZXNzYWdlLiIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICJkZWZhdWx0IjogMQogICAgICAgIH0sCiAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEaXNhYmxlcyBhY2tub3dsZWRnaW5nIG1lc3NhZ2VzIHRoYXQgYXJlIHJlY2VpdmVkIGJ5IHRoZSBTdHJlYW0uIiwKICAgICAgICAgICJ0eXBlIjogImJvb2xlYW4iLAogICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgIH0sCiAgICAgICAgInRlbXBsYXRlX293bmVyIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICB9LAogICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gYSBTdHJlYW0gcmVhY2ggaXQncyBsaW1pdHMgZWl0aGVyIG9sZCBtZXNzYWdlcyBhcmUgZGVsZXRlZCBvciBuZXcgb25lcyBhcmUgZGVuaWVkIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgIm9sZCIsCiAgICAgICAgICAgICJuZXciCiAgICAgICAgICBdLAogICAgICAgICAgImRlZmF1bHQiOiAib2xkIgogICAgICAgIH0sCiAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgd2luZG93IHRvIHRyYWNrIGR1cGxpY2F0ZSBtZXNzYWdlcyBmb3IsIGV4cHJlc3NlZCBpbiBtaWNyb3NlY29uZHMuIDAgZm9yIGRlZmF1bHQiLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAiZGVmYXVsdCI6IDAKICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdCn0K")
-	schemas["io.nats.jetstream.api.v1.stream_create_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9jcmVhdGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLkNSRUFURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX2NyZWF0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdGF0ZSIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICJhbGxPZiI6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAibWVtb3J5IgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgICAgICAgICAibmV3IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfQogICAgICAgICAgXQogICAgICAgIH0sCiAgICAgICAgInN0YXRlIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEZXRhaWwgYWJvdXQgdGhlIGN1cnJlbnQgU3RhdGUgb2YgdGhlIFN0cmVhbSIsCiAgICAgICAgICAiYWxsT2YiOiBbCiAgICAgICAgICAgIHsKICAgICAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICJtZXNzYWdlcyIsCiAgICAgICAgICAgICAgICAiYnl0ZXMiLAogICAgICAgICAgICAgICAgImZpcnN0X3NlcSIsCiAgICAgICAgICAgICAgICAibGFzdF9zZXEiLAogICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAibWVzc2FnZXMiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTnVtYmVyIG9mIG1lc3NhZ2VzIHN0b3JlZCBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImJ5dGVzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkNvbWJpbmVkIHNpemUgb2YgYWxsIG1lc3NhZ2VzIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZmlyc3Rfc2VxIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlNlcXVlbmNlIG51bWJlciBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImZpcnN0X3RzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWVzdGFtcCBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImxhc3Rfc2VxIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlNlcXVlbmNlIG51bWJlciBvZiB0aGUgbGFzdCBtZXNzYWdlIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibGFzdF90cyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lc3RhbXAgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBDb25zdW1lcnMgYXR0YWNoZWQgdG8gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfQogICAgICAgICAgXQogICAgICAgIH0KICAgICAgfQogICAgfSwKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9CiAgXSwKICAicHJvcGVydGllcyI6IHsKICAgICJ0eXBlIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiY29uc3QiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV9jcmVhdGVfcmVzcG9uc2UiCiAgICB9CiAgfQp9Cg==")
+	schemas["io.nats.jetstream.api.v1.stream_create_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9jcmVhdGVfcmVxdWVzdC5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXF1ZXN0IHRvIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5TVFJFQU0uQ1JFQVRFIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fY3JlYXRlX3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgImFsbE9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAicmV0ZW50aW9uIiwKICAgICAgICAibWF4X2NvbnN1bWVycyIsCiAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAibWF4X2J5dGVzIiwKICAgICAgICAibWF4X2FnZSIsCiAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICJudW1fcmVwbGljYXMiCiAgICAgIF0sCiAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAibmFtZSI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgIH0sCiAgICAgICAgInN1YmplY3RzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICJ0eXBlIjogImFycmF5IiwKICAgICAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgICAgfQogICAgICAgIH0sCiAgICAgICAgInJldGVudGlvbiI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICJpbnRlcmVzdCIsCiAgICAgICAgICAgICJ3b3JrcXVldWUiCiAgICAgICAgICBdLAogICAgICAgICAgImRlZmF1bHQiOiAibGltaXRzIgogICAgICAgIH0sCiAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgQ29uc3VtZXJzIGNhbiBiZSBkZWZpbmVkIGZvciBhIGdpdmVuIFN0cmVhbS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgIH0sCiAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IG1lc3NhZ2VzIG1heSBiZSBpbiBhIFN0cmVhbSwgb2xkZXN0IG1lc3NhZ2VzIHdpbGwgYmUgcmVtb3ZlZCBpZiB0aGUgU3RyZWFtIGV4Y2VlZHMgdGhpcyBzaXplLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgfSwKICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBiaWcgdGhlIFN0cmVhbSBtYXkgYmUsIHdoZW4gdGhlIGNvbWJpbmVkIHN0cmVhbSBzaXplIGV4Y2VlZHMgdGhpcyBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICB9LAogICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIk1heGltdW0gYWdlIG9mIGFueSBtZXNzYWdlIGluIHRoZSBzdHJlYW0sIGV4cHJlc3NlZCBpbiBtaWNyb3NlY29uZHMuIDAgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAiZGVmYXVsdCI6IDAKICAgICAgICB9LAogICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGxhcmdlc3QgbWVzc2FnZSB0aGF0IHdpbGwgYmUgYWNjZXB0ZWQgYnkgdGhlIFN0cmVhbS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgIH0sCiAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHN0b3JhZ2UgYmFja2VuZCB0byB1c2UgZm9yIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgImZpbGUiLAogICAgICAgICAgICAibWVtb3J5IgogICAgICAgICAgXSwKICAgICAgICAgICJkZWZhdWx0IjogImZpbGUiCiAgICAgICAgfSwKICAgICAgICAibnVtX3JlcGxpY2FzIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IHJlcGxpY2FzIHRvIGtlZXAgZm9yIGVhY2ggbWVzc2FnZS4iLAogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDEsCiAgICAgICAgICAiZGVmYXVsdCI6IDEKICAgICAgICB9LAogICAgICAgICJub19hY2siOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGlzYWJsZXMgYWNrbm93bGVkZ2luZyBtZXNzYWdlcyB0aGF0IGFyZSByZWNlaXZlZCBieSB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgICAgICJkZWZhdWx0IjogZmFsc2UKICAgICAgICB9LAogICAgICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIHRoZSBTdHJlYW0gaXMgbWFuYWdlZCBieSBhIFN0cmVhbSBUZW1wbGF0ZSB0aGlzIGlkZW50aWZpZXMgdGhlIHRlbXBsYXRlIHRoYXQgbWFuYWdlcyB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgfSwKICAgICAgICAiZGlzY2FyZCI6IHsKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIGEgU3RyZWFtIHJlYWNoIGl0J3MgbGltaXRzIGVpdGhlciBvbGQgbWVzc2FnZXMgYXJlIGRlbGV0ZWQgb3IgbmV3IG9uZXMgYXJlIGRlbmllZCIsCiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICJvbGQiLAogICAgICAgICAgICAibmV3IgogICAgICAgICAgXSwKICAgICAgICAgICJkZWZhdWx0IjogIm9sZCIKICAgICAgICB9LAogICAgICAgICJkdXBsaWNhdGVfd2luZG93IjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHdpbmRvdyB0byB0cmFjayBkdXBsaWNhdGUgbWVzc2FnZXMgZm9yLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciBkZWZhdWx0IiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgfQogICAgICB9CiAgICB9CiAgXQp9Cg==")
+	schemas["io.nats.jetstream.api.v1.stream_create_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9jcmVhdGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLkNSRUFURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX2NyZWF0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdGF0ZSIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICJhbGxPZiI6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAiZmlsZSIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibnVtX3JlcGxpY2FzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgcmVwbGljYXMgdG8ga2VlcCBmb3IgZWFjaCBtZXNzYWdlLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJub19hY2siOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEaXNhYmxlcyBhY2tub3dsZWRnaW5nIG1lc3NhZ2VzIHRoYXQgYXJlIHJlY2VpdmVkIGJ5IHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogZmFsc2UKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAidGVtcGxhdGVfb3duZXIiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIHRoZSBTdHJlYW0gaXMgbWFuYWdlZCBieSBhIFN0cmVhbSBUZW1wbGF0ZSB0aGlzIGlkZW50aWZpZXMgdGhlIHRlbXBsYXRlIHRoYXQgbWFuYWdlcyB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZGlzY2FyZCI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gYSBTdHJlYW0gcmVhY2ggaXQncyBsaW1pdHMgZWl0aGVyIG9sZCBtZXNzYWdlcyBhcmUgZGVsZXRlZCBvciBuZXcgb25lcyBhcmUgZGVuaWVkIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgIm9sZCIsCiAgICAgICAgICAgICAgICAgICAgIm5ldyIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAib2xkIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkdXBsaWNhdGVfd2luZG93IjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgd2luZG93IHRvIHRyYWNrIGR1cGxpY2F0ZSBtZXNzYWdlcyBmb3IsIGV4cHJlc3NlZCBpbiBtaWNyb3NlY29uZHMuIDAgZm9yIGRlZmF1bHQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IDAKICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0KICAgICAgICAgIF0KICAgICAgICB9LAogICAgICAgICJzdGF0ZSI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGV0YWlsIGFib3V0IHRoZSBjdXJyZW50IFN0YXRlIG9mIHRoZSBTdHJlYW0iLAogICAgICAgICAgImFsbE9mIjogWwogICAgICAgICAgICB7CiAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAibWVzc2FnZXMiLAogICAgICAgICAgICAgICAgImJ5dGVzIiwKICAgICAgICAgICAgICAgICJmaXJzdF9zZXEiLAogICAgICAgICAgICAgICAgImxhc3Rfc2VxIiwKICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm1lc3NhZ2VzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBtZXNzYWdlcyBzdG9yZWQgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJieXRlcyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJDb21iaW5lZCBzaXplIG9mIGFsbCBtZXNzYWdlcyBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImZpcnN0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGZpcnN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJmaXJzdF90cyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lc3RhbXAgb2YgdGhlIGZpcnN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJsYXN0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImxhc3RfdHMiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZXN0YW1wIG9mIHRoZSBsYXN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJOdW1iZXIgb2YgQ29uc3VtZXJzIGF0dGFjaGVkIHRvIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0KICAgICAgICAgIF0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiZXJyb3IiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJlcnJvciI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJjb2RlIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSFRUUCBsaWtlIGVycm9yIGNvZGUgaW4gdGhlIDMwMCB0byA1MDAgcmFuZ2UiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgICAgICJtYXhpbXVtIjogNjk5CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGh1bWFuIGZyaWVuZGx5IGRlc2NyaXB0aW9uIG9mIHRoZSBlcnJvciIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fY3JlYXRlX3Jlc3BvbnNlIgogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.api.v1.stream_delete_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9kZWxldGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLkRFTEVURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX2RlbGV0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgewogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgInN1Y2Nlc3MiCiAgICAgIF0sCiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJzdWNjZXNzIjogewogICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIKICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX2RlbGV0ZV9yZXNwb25zZSIKICAgIH0KICB9Cn0K")
-	schemas["io.nats.jetstream.api.v1.stream_info_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9pbmZvX3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5JTkZPIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1faW5mb19yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdGF0ZSIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICJhbGxPZiI6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAibWVtb3J5IgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgICAgICAgICAibmV3IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfQogICAgICAgICAgXQogICAgICAgIH0sCiAgICAgICAgInN0YXRlIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEZXRhaWwgYWJvdXQgdGhlIGN1cnJlbnQgU3RhdGUgb2YgdGhlIFN0cmVhbSIsCiAgICAgICAgICAiYWxsT2YiOiBbCiAgICAgICAgICAgIHsKICAgICAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICJtZXNzYWdlcyIsCiAgICAgICAgICAgICAgICAiYnl0ZXMiLAogICAgICAgICAgICAgICAgImZpcnN0X3NlcSIsCiAgICAgICAgICAgICAgICAibGFzdF9zZXEiLAogICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAibWVzc2FnZXMiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTnVtYmVyIG9mIG1lc3NhZ2VzIHN0b3JlZCBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImJ5dGVzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkNvbWJpbmVkIHNpemUgb2YgYWxsIG1lc3NhZ2VzIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZmlyc3Rfc2VxIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlNlcXVlbmNlIG51bWJlciBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImZpcnN0X3RzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWVzdGFtcCBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImxhc3Rfc2VxIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlNlcXVlbmNlIG51bWJlciBvZiB0aGUgbGFzdCBtZXNzYWdlIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibGFzdF90cyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lc3RhbXAgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBDb25zdW1lcnMgYXR0YWNoZWQgdG8gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfQogICAgICAgICAgXQogICAgICAgIH0KICAgICAgfQogICAgfSwKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9CiAgXSwKICAicHJvcGVydGllcyI6IHsKICAgICJ0eXBlIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiY29uc3QiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV9pbmZvX3Jlc3BvbnNlIgogICAgfQogIH0KfQo=")
+	schemas["io.nats.jetstream.api.v1.stream_info_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9pbmZvX3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5JTkZPIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1faW5mb19yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdGF0ZSIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICJhbGxPZiI6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAiZmlsZSIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibnVtX3JlcGxpY2FzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgcmVwbGljYXMgdG8ga2VlcCBmb3IgZWFjaCBtZXNzYWdlLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJub19hY2siOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEaXNhYmxlcyBhY2tub3dsZWRnaW5nIG1lc3NhZ2VzIHRoYXQgYXJlIHJlY2VpdmVkIGJ5IHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogZmFsc2UKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAidGVtcGxhdGVfb3duZXIiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIHRoZSBTdHJlYW0gaXMgbWFuYWdlZCBieSBhIFN0cmVhbSBUZW1wbGF0ZSB0aGlzIGlkZW50aWZpZXMgdGhlIHRlbXBsYXRlIHRoYXQgbWFuYWdlcyB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZGlzY2FyZCI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gYSBTdHJlYW0gcmVhY2ggaXQncyBsaW1pdHMgZWl0aGVyIG9sZCBtZXNzYWdlcyBhcmUgZGVsZXRlZCBvciBuZXcgb25lcyBhcmUgZGVuaWVkIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgIm9sZCIsCiAgICAgICAgICAgICAgICAgICAgIm5ldyIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAib2xkIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkdXBsaWNhdGVfd2luZG93IjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgd2luZG93IHRvIHRyYWNrIGR1cGxpY2F0ZSBtZXNzYWdlcyBmb3IsIGV4cHJlc3NlZCBpbiBtaWNyb3NlY29uZHMuIDAgZm9yIGRlZmF1bHQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IDAKICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0KICAgICAgICAgIF0KICAgICAgICB9LAogICAgICAgICJzdGF0ZSI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGV0YWlsIGFib3V0IHRoZSBjdXJyZW50IFN0YXRlIG9mIHRoZSBTdHJlYW0iLAogICAgICAgICAgImFsbE9mIjogWwogICAgICAgICAgICB7CiAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAibWVzc2FnZXMiLAogICAgICAgICAgICAgICAgImJ5dGVzIiwKICAgICAgICAgICAgICAgICJmaXJzdF9zZXEiLAogICAgICAgICAgICAgICAgImxhc3Rfc2VxIiwKICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm1lc3NhZ2VzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBtZXNzYWdlcyBzdG9yZWQgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJieXRlcyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJDb21iaW5lZCBzaXplIG9mIGFsbCBtZXNzYWdlcyBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImZpcnN0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGZpcnN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJmaXJzdF90cyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lc3RhbXAgb2YgdGhlIGZpcnN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJsYXN0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImxhc3RfdHMiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZXN0YW1wIG9mIHRoZSBsYXN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJOdW1iZXIgb2YgQ29uc3VtZXJzIGF0dGFjaGVkIHRvIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0KICAgICAgICAgIF0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiZXJyb3IiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJlcnJvciI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJjb2RlIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSFRUUCBsaWtlIGVycm9yIGNvZGUgaW4gdGhlIDMwMCB0byA1MDAgcmFuZ2UiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgICAgICJtYXhpbXVtIjogNjk5CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGh1bWFuIGZyaWVuZGx5IGRlc2NyaXB0aW9uIG9mIHRoZSBlcnJvciIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1faW5mb19yZXNwb25zZSIKICAgIH0KICB9Cn0K")
 	schemas["io.nats.jetstream.api.v1.stream_list_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9saXN0X3JlcXVlc3QuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVxdWVzdCB0byB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLkxJU1QgQVBJIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV9saXN0X3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgImFsbE9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJvZmZzZXQiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJvZmZzZXQiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIF0KfQo=")
-	schemas["io.nats.jetstream.api.v1.stream_list_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9saXN0X3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5MSVNUIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fbGlzdF9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJhbGxPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJ0b3RhbCIsCiAgICAgICAgIm9mZnNldCIsCiAgICAgICAgImxpbWl0IgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAidG90YWwiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0sCiAgICAgICAgIm9mZnNldCI6IHsKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgfSwKICAgICAgICAibGltaXQiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgIm9uZU9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgImVycm9yIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiZXJyb3IiOiB7CiAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29kZSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImNvZGUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDMwMCwKICAgICAgICAgICAgICAibWF4aW11bSI6IDY5OQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGVzY3JpcHRpb24iOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAic3RyZWFtcyIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbnN1bWVycyI6IHsKICAgICAgICAgICJ0eXBlIjogImFycmF5IiwKICAgICAgICAgICJpdGVtcyI6IHsKICAgICAgICAgICAgImFsbE9mIjogWwogICAgICAgICAgICAgIHsKICAgICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAgICJjb25maWciLAogICAgICAgICAgICAgICAgICAic3RhdGUiCiAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAgICJjb25maWciOiB7CiAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAiYWxsT2YiOiBbCiAgICAgICAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICAgICAgICAgICAic3ViamVjdHMiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfY29uc3VtZXJzIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAibWF4X21zZ3MiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfYWdlIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAic3RvcmFnZSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgICAgICAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgICAgICAgICAgICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInBhdHRlcm4iOiAiXlteLio+XSokIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAwCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAic3ViamVjdHMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxCiAgICAgICAgICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAicmV0ZW50aW9uIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAibGltaXRzIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIndvcmtxdWV1ZSIKICAgICAgICAgICAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAibWF4X21zZ3MiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm1heF9hZ2UiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAic3RvcmFnZSI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtZW1vcnkiCiAgICAgICAgICAgICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAibWVtb3J5IgogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAxLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAibm9fYWNrIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IGZhbHNlCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAidGVtcGxhdGVfb3duZXIiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiB0aGUgU3RyZWFtIGlzIG1hbmFnZWQgYnkgYSBTdHJlYW0gVGVtcGxhdGUgdGhpcyBpZGVudGlmaWVzIHRoZSB0ZW1wbGF0ZSB0aGF0IG1hbmFnZXMgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImRpc2NhcmQiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIm5ldyIKICAgICAgICAgICAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAiZHVwbGljYXRlX3dpbmRvdyI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgInN0YXRlIjogewogICAgICAgICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRldGFpbCBhYm91dCB0aGUgY3VycmVudCBTdGF0ZSBvZiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAiYWxsT2YiOiBbCiAgICAgICAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm1lc3NhZ2VzIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAiYnl0ZXMiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJmaXJzdF9zZXEiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJsYXN0X3NlcSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IgogICAgICAgICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAibWVzc2FnZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJOdW1iZXIgb2YgbWVzc2FnZXMgc3RvcmVkIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAiYnl0ZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJDb21iaW5lZCBzaXplIG9mIGFsbCBtZXNzYWdlcyBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImZpcnN0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlNlcXVlbmNlIG51bWJlciBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImZpcnN0X3RzIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZXN0YW1wIG9mIHRoZSBmaXJzdCBtZXNzYWdlIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAibGFzdF9zZXEiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImxhc3RfdHMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lc3RhbXAgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTnVtYmVyIG9mIENvbnN1bWVycyBhdHRhY2hlZCB0byB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgICAgIF0KICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgXQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fbGlzdF9yZXNwb25zZSIKICAgIH0KICB9Cn0K")
+	schemas["io.nats.jetstream.api.v1.stream_list_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9saXN0X3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5MSVNUIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fbGlzdF9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJhbGxPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJ0b3RhbCIsCiAgICAgICAgIm9mZnNldCIsCiAgICAgICAgImxpbWl0IgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAidG90YWwiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0sCiAgICAgICAgIm9mZnNldCI6IHsKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgfSwKICAgICAgICAibGltaXQiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgIm9uZU9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgImVycm9yIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiZXJyb3IiOiB7CiAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29kZSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImNvZGUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDMwMCwKICAgICAgICAgICAgICAibWF4aW11bSI6IDY5OQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGVzY3JpcHRpb24iOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAic3RyZWFtcyIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbnN1bWVycyI6IHsKICAgICAgICAgICJ0eXBlIjogImFycmF5IiwKICAgICAgICAgICJpdGVtcyI6IHsKICAgICAgICAgICAgImFsbE9mIjogWwogICAgICAgICAgICAgIHsKICAgICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAgICJjb25maWciLAogICAgICAgICAgICAgICAgICAic3RhdGUiCiAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAgICJjb25maWciOiB7CiAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAiYWxsT2YiOiBbCiAgICAgICAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICAgICAgICAgICAic3ViamVjdHMiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfY29uc3VtZXJzIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAibWF4X21zZ3MiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfYWdlIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAic3RvcmFnZSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgICAgICAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgICAgICAgICAgICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInBhdHRlcm4iOiAiXlteLio+XSokIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAwCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAic3ViamVjdHMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxCiAgICAgICAgICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAicmV0ZW50aW9uIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAibGltaXRzIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIndvcmtxdWV1ZSIKICAgICAgICAgICAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAibWF4X21zZ3MiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm1heF9hZ2UiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICAgICAgICAic3RvcmFnZSI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtZW1vcnkiCiAgICAgICAgICAgICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAiZmlsZSIKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgcmVwbGljYXMgdG8ga2VlcCBmb3IgZWFjaCBtZXNzYWdlLiIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMQogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEaXNhYmxlcyBhY2tub3dsZWRnaW5nIG1lc3NhZ2VzIHRoYXQgYXJlIHJlY2VpdmVkIGJ5IHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogImJvb2xlYW4iLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgInRlbXBsYXRlX293bmVyIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gYSBTdHJlYW0gcmVhY2ggaXQncyBsaW1pdHMgZWl0aGVyIG9sZCBtZXNzYWdlcyBhcmUgZGVsZXRlZCBvciBuZXcgb25lcyBhcmUgZGVuaWVkIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIm9sZCIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJuZXciCiAgICAgICAgICAgICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAib2xkIgogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgd2luZG93IHRvIHRyYWNrIGR1cGxpY2F0ZSBtZXNzYWdlcyBmb3IsIGV4cHJlc3NlZCBpbiBtaWNyb3NlY29uZHMuIDAgZm9yIGRlZmF1bHQiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IDAKICAgICAgICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgICAgICBdCiAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICJzdGF0ZSI6IHsKICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEZXRhaWwgYWJvdXQgdGhlIGN1cnJlbnQgU3RhdGUgb2YgdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICAgImFsbE9mIjogWwogICAgICAgICAgICAgICAgICAgICAgewogICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgICAgICAgICAgICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgICAgICAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgICAgICAgICAgICJtZXNzYWdlcyIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgImJ5dGVzIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAiZmlyc3Rfc2VxIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAibGFzdF9zZXEiLAogICAgICAgICAgICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCIKICAgICAgICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgIm1lc3NhZ2VzIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTnVtYmVyIG9mIG1lc3NhZ2VzIHN0b3JlZCBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImJ5dGVzIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQ29tYmluZWQgc2l6ZSBvZiBhbGwgbWVzc2FnZXMgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJmaXJzdF9zZXEiOiB7CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGZpcnN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJmaXJzdF90cyI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWVzdGFtcCBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICAgImxhc3Rfc2VxIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiU2VxdWVuY2UgbnVtYmVyIG9mIHRoZSBsYXN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJsYXN0X3RzIjogewogICAgICAgICAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZXN0YW1wIG9mIHRoZSBsYXN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZm9ybWF0IjogImRhdGUtdGltZSIKICAgICAgICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCI6IHsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBDb25zdW1lcnMgYXR0YWNoZWQgdG8gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgICAgICBdCiAgICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIF0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX2xpc3RfcmVzcG9uc2UiCiAgICB9CiAgfQp9Cg==")
 	schemas["io.nats.jetstream.api.v1.stream_msg_delete_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9tc2dfZGVsZXRlX3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5NU0cuREVMRVRFIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fbXNnX2RlbGV0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgInN1Y2Nlc3MiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJzdWNjZXNzIjogewogICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIKICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX21zZ19kZWxldGVfcmVzcG9uc2UiCiAgICB9CiAgfQp9Cg==")
 	schemas["io.nats.jetstream.api.v1.stream_msg_get_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9tc2dfZ2V0X3JlcXVlc3QuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVxdWVzdCB0byB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLk1TRy5HRVQgQVBJIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV9tc2dfZ2V0X3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgInJlcXVpcmVkIjogWwogICAgInNlcSIKICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInNlcSI6IHsKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJTdHJlYW0gc2VxdWVuY2UgbnVtYmVyIG9mIHRoZSBtZXNzYWdlIHRvIHJldHJpZXZlIgogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.api.v1.stream_msg_get_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9tc2dfZ2V0X3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5NU0cuR0VUIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fbXNnX2dldF9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgIm1lc3NhZ2UiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJtZXNzYWdlIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAic3ViamVjdCIsCiAgICAgICAgICAgICJzZXEiLAogICAgICAgICAgICAiZGF0YSIsCiAgICAgICAgICAgICJ0aW1lIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAic3ViamVjdCI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3ViamVjdCB0aGUgbWVzc2FnZSB3YXMgb3JpZ2luYWxseSByZWNlaXZlZCBvbiIsCiAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgInNlcSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHNlcXVlbmNlIG51bWJlciBvZiB0aGUgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRhdGEiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGJhc2U2NCBlbmNvZGVkIHBheWxvYWQgb2YgdGhlIG1lc3NhZ2UgYm9keSIsCiAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDAKICAgICAgICAgICAgfSwKICAgICAgICAgICAgInRpbWUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgdGhlIG1lc3NhZ2Ugd2FzIHJlY2VpdmVkIiwKICAgICAgICAgICAgICAiZm9ybWF0IjogImRhdGUtdGltZSIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fbXNnX2dldF9yZXNwb25zZSIKICAgIH0KICB9Cn0K")
@@ -975,12 +975,12 @@ func init() {
 	schemas["io.nats.jetstream.api.v1.stream_snapshot_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9zbmFwc2hvdF9yZXNwb25zZS5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXNwb25zZSBmcm9tIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5TVFJFQU0uU05BUFNIT1QgQVBJIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV9zbmFwc2hvdF9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAib25lT2YiOiBbCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiZXJyb3IiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJlcnJvciI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJjb2RlIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSFRUUCBsaWtlIGVycm9yIGNvZGUgaW4gdGhlIDMwMCB0byA1MDAgcmFuZ2UiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgICAgICJtYXhpbXVtIjogNjk5CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGh1bWFuIGZyaWVuZGx5IGRlc2NyaXB0aW9uIG9mIHRoZSBlcnJvciIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfSwKICAgIHsKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJudW1fYmxrcyIsCiAgICAgICAgImJsa19zaXplIgogICAgICBdLAogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAibnVtX2Jsa3MiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgYXBwcm94aW1hdGUgbnVtYmVyIG9mIGJsb2NrcyBpbiB0aGUgc25hcHNob3QiLAogICAgICAgICAgIm1pbmltdW0iOiAxCiAgICAgICAgfSwKICAgICAgICAiYmxrX3NpemUiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHNpemUgb2YgZWFjaCBibG9jayBiZWluZyBzZW50IiwKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAxCiAgICAgICAgfQogICAgICB9CiAgICB9CiAgXSwKICAicHJvcGVydGllcyI6IHsKICAgICJ0eXBlIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiY29uc3QiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV9zbmFwc2hvdF9yZXNwb25zZSIKICAgIH0KICB9Cn0K")
 	schemas["io.nats.jetstream.api.v1.stream_snapshot_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9zbmFwc2hvdF9yZXF1ZXN0Lmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlcXVlc3QgdG8gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5TTkFQU0hPVCBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3NuYXBzaG90X3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgInJlcXVpcmVkIjogWwogICAgImRlbGl2ZXJfc3ViamVjdCIKICBdLAogICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICJwcm9wZXJ0aWVzIjogewogICAgImRlbGl2ZXJfc3ViamVjdCI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBOQVRTIHN1YmplY3Qgd2hlcmUgdGhlIHNuYXBzaG90IHdpbGwgYmUgZGVsaXZlcmVkIiwKICAgICAgIm1pbkxlbmd0aCI6IDEKICAgIH0sCiAgICAibm9fY29uc3VtZXJzIjogewogICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdHJ1ZSBjb25zdW1lciBzdGF0ZXMgYW5kIGNvbmZpZ3VyYXRpb25zIHdpbGwgbm90IGJlIHByZXNlbnQgaW4gdGhlIHNuYXBzaG90IgogICAgfSwKICAgICJjaHVua19zaXplIjogewogICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzaXplIG9mIGRhdGEgY2h1bmtzIHRvIHNlbmQgdG8gZGVsaXZlcl9zdWJqZWN0IiwKICAgICAgIm1pbmltdW0iOiAxMDI0CiAgICB9LAogICAgImpzY2siOiB7CiAgICAgICJ0eXBlIjogImJvb2xlYW4iLAogICAgICAiZGVzY3JpcHRpb24iOiAiQ2hlY2sgYWxsIG1lc3NhZ2UncyBjaGVja3N1bXMgcHJpb3IgdG8gc25hcHNob3QiLAogICAgICAiZGVmYXVsdCI6IGZhbHNlCiAgICB9CiAgfQp9Cg==")
 	schemas["io.nats.jetstream.api.v1.stream_restore_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV9yZXN0b3JlX3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5SRVNUT1JFIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fcmVzdG9yZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAib25lT2YiOiBbCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiZXJyb3IiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJlcnJvciI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJjb2RlIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSFRUUCBsaWtlIGVycm9yIGNvZGUgaW4gdGhlIDMwMCB0byA1MDAgcmFuZ2UiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgICAgICJtYXhpbXVtIjogNjk5CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGh1bWFuIGZyaWVuZGx5IGRlc2NyaXB0aW9uIG9mIHRoZSBlcnJvciIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfSwKICAgIHsKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJkZWxpdmVyX3N1YmplY3QiCiAgICAgIF0sCiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJkZWxpdmVyX3N1YmplY3QiOiB7CiAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBTdWJqZWN0IHRvIHNlbmQgcmVzdG9yZSBjaHVua3MgdG8iLAogICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3Jlc3RvcmVfcmVzcG9uc2UiCiAgICB9CiAgfQp9Cg==")
-	schemas["io.nats.jetstream.api.v1.stream_template_create_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9jcmVhdGVfcmVxdWVzdC5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXF1ZXN0IHRvIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5TVFJFQU0uVEVNUExBVEUuQ1JFQVRFIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfY3JlYXRlX3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgImFsbE9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJuYW1lIiwKICAgICAgICAiY29uZmlnIiwKICAgICAgICAibWF4X3N0cmVhbXMiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJuYW1lIjogewogICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgVGVtcGxhdGUiCiAgICAgICAgfSwKICAgICAgICAiY29uZmlnIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0ZW1wbGF0ZSBjb25maWd1cmF0aW9uIHRvIGNyZWF0ZSBTdHJlYW1zIHdpdGgiLAogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgICAgICJtYXhfY29uc3VtZXJzIiwKICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgICAgICJtYXhfYWdlIiwKICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICAgICAgXSwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgInBhdHRlcm4iOiAiXlteLio+XSokIiwKICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICB9LAogICAgICAgICAgICAic3ViamVjdHMiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEsCiAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxCiAgICAgICAgICAgICAgfQogICAgICAgICAgICB9LAogICAgICAgICAgICAicmV0ZW50aW9uIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAgICAgIndvcmtxdWV1ZSIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgICAgICJtZW1vcnkiCiAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJtZW1vcnkiCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IHJlcGxpY2FzIHRvIGtlZXAgZm9yIGVhY2ggbWVzc2FnZS4iLAogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICAgICAiZGVmYXVsdCI6IDEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGlzYWJsZXMgYWNrbm93bGVkZ2luZyBtZXNzYWdlcyB0aGF0IGFyZSByZWNlaXZlZCBieSB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgICAgICB9LAogICAgICAgICAgICAidGVtcGxhdGVfb3duZXIiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIGEgU3RyZWFtIHJlYWNoIGl0J3MgbGltaXRzIGVpdGhlciBvbGQgbWVzc2FnZXMgYXJlIGRlbGV0ZWQgb3IgbmV3IG9uZXMgYXJlIGRlbmllZCIsCiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICJvbGQiLAogICAgICAgICAgICAgICAgIm5ldyIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJkZWZhdWx0IjogIm9sZCIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHdpbmRvdyB0byB0cmFjayBkdXBsaWNhdGUgbWVzc2FnZXMgZm9yLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciBkZWZhdWx0IiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9LAogICAgICAgICJtYXhfc3RyZWFtcyI6IHsKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBtYXhpbXVtIG51bWJlciBvZiBzdHJlYW1zIHRvIGFsbG93IHVzaW5nIHRoaXMgVGVtcGxhdGUiLAogICAgICAgICAgIm1pbmltdW0iOiAtMQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0KfQo=")
-	schemas["io.nats.jetstream.api.v1.stream_template_create_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9jcmVhdGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLlRFTVBMQVRFLkNSRUFURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3RlbXBsYXRlX2NyZWF0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdHJlYW1zIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiY29uZmlnIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAibmFtZSIsCiAgICAgICAgICAgICJjb25maWciLAogICAgICAgICAgICAibWF4X3N0cmVhbXMiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBUZW1wbGF0ZSIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRlbXBsYXRlIGNvbmZpZ3VyYXRpb24gdG8gY3JlYXRlIFN0cmVhbXMgd2l0aCIsCiAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAic3ViamVjdHMiLAogICAgICAgICAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyIsCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiLAogICAgICAgICAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgICAgICAgICAibWF4X2FnZSIsCiAgICAgICAgICAgICAgICAic3RvcmFnZSIsCiAgICAgICAgICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAibmFtZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN1YmplY3RzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAgICAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInJldGVudGlvbiI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAibGltaXRzIiwKICAgICAgICAgICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAgICAgICAgICJ3b3JrcXVldWUiCiAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9hZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAic3RvcmFnZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgICAgICAgICAibWVtb3J5IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJtZW1vcnkiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IHJlcGxpY2FzIHRvIGtlZXAgZm9yIGVhY2ggbWVzc2FnZS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAxLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IDEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibm9fYWNrIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGlzYWJsZXMgYWNrbm93bGVkZ2luZyBtZXNzYWdlcyB0aGF0IGFyZSByZWNlaXZlZCBieSB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImJvb2xlYW4iLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IGZhbHNlCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInRlbXBsYXRlX293bmVyIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiB0aGUgU3RyZWFtIGlzIG1hbmFnZWQgYnkgYSBTdHJlYW0gVGVtcGxhdGUgdGhpcyBpZGVudGlmaWVzIHRoZSB0ZW1wbGF0ZSB0aGF0IG1hbmFnZXMgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImRpc2NhcmQiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIGEgU3RyZWFtIHJlYWNoIGl0J3MgbGltaXRzIGVpdGhlciBvbGQgbWVzc2FnZXMgYXJlIGRlbGV0ZWQgb3IgbmV3IG9uZXMgYXJlIGRlbmllZCIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICJvbGQiLAogICAgICAgICAgICAgICAgICAgICJuZXciCiAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogIm9sZCIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZHVwbGljYXRlX3dpbmRvdyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHdpbmRvdyB0byB0cmFjayBkdXBsaWNhdGUgbWVzc2FnZXMgZm9yLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciBkZWZhdWx0IiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgfQogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X3N0cmVhbXMiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBtYXhpbXVtIG51bWJlciBvZiBzdHJlYW1zIHRvIGFsbG93IHVzaW5nIHRoaXMgVGVtcGxhdGUiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogLTEKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0sCiAgICAgICAgInN0cmVhbXMiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTGlzdCBvZiBTdHJlYW1zIG1hbmFnZWQgYnkgdGhpcyBUZW1wbGF0ZSIsCiAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiZXJyb3IiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJlcnJvciI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJjb2RlIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSFRUUCBsaWtlIGVycm9yIGNvZGUgaW4gdGhlIDMwMCB0byA1MDAgcmFuZ2UiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgICAgICJtYXhpbXVtIjogNjk5CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGh1bWFuIGZyaWVuZGx5IGRlc2NyaXB0aW9uIG9mIHRoZSBlcnJvciIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfY3JlYXRlX3Jlc3BvbnNlIgogICAgfQogIH0KfQo=")
+	schemas["io.nats.jetstream.api.v1.stream_template_create_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9jcmVhdGVfcmVxdWVzdC5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXF1ZXN0IHRvIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5TVFJFQU0uVEVNUExBVEUuQ1JFQVRFIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfY3JlYXRlX3JlcXVlc3QiLAogICJ0eXBlIjogIm9iamVjdCIsCiAgImFsbE9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJuYW1lIiwKICAgICAgICAiY29uZmlnIiwKICAgICAgICAibWF4X3N0cmVhbXMiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJuYW1lIjogewogICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgVGVtcGxhdGUiCiAgICAgICAgfSwKICAgICAgICAiY29uZmlnIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0ZW1wbGF0ZSBjb25maWd1cmF0aW9uIHRvIGNyZWF0ZSBTdHJlYW1zIHdpdGgiLAogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgICAgICJtYXhfY29uc3VtZXJzIiwKICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgICAgICJtYXhfYWdlIiwKICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICAgICAgXSwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgInBhdHRlcm4iOiAiXlteLio+XSokIiwKICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICB9LAogICAgICAgICAgICAic3ViamVjdHMiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEsCiAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxCiAgICAgICAgICAgICAgfQogICAgICAgICAgICB9LAogICAgICAgICAgICAicmV0ZW50aW9uIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAgICAgIndvcmtxdWV1ZSIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgfSwKICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgICAgICJtZW1vcnkiCiAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJmaWxlIgogICAgICAgICAgICB9LAogICAgICAgICAgICAibnVtX3JlcGxpY2FzIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDEsCiAgICAgICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJub19hY2siOiB7CiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICJ0eXBlIjogImJvb2xlYW4iLAogICAgICAgICAgICAgICJkZWZhdWx0IjogZmFsc2UKICAgICAgICAgICAgfSwKICAgICAgICAgICAgInRlbXBsYXRlX293bmVyIjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIHRoZSBTdHJlYW0gaXMgbWFuYWdlZCBieSBhIFN0cmVhbSBUZW1wbGF0ZSB0aGlzIGlkZW50aWZpZXMgdGhlIHRlbXBsYXRlIHRoYXQgbWFuYWdlcyB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGlzY2FyZCI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgICAgICJuZXciCiAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkdXBsaWNhdGVfd2luZG93IjogewogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfSwKICAgICAgICAibWF4X3N0cmVhbXMiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbWF4aW11bSBudW1iZXIgb2Ygc3RyZWFtcyB0byBhbGxvdyB1c2luZyB0aGlzIFRlbXBsYXRlIiwKICAgICAgICAgICJtaW5pbXVtIjogLTEKICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdCn0K")
+	schemas["io.nats.jetstream.api.v1.stream_template_create_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9jcmVhdGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLlRFTVBMQVRFLkNSRUFURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3RlbXBsYXRlX2NyZWF0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdHJlYW1zIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiY29uZmlnIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAibmFtZSIsCiAgICAgICAgICAgICJjb25maWciLAogICAgICAgICAgICAibWF4X3N0cmVhbXMiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBUZW1wbGF0ZSIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRlbXBsYXRlIGNvbmZpZ3VyYXRpb24gdG8gY3JlYXRlIFN0cmVhbXMgd2l0aCIsCiAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAic3ViamVjdHMiLAogICAgICAgICAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyIsCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiLAogICAgICAgICAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgICAgICAgICAibWF4X2FnZSIsCiAgICAgICAgICAgICAgICAic3RvcmFnZSIsCiAgICAgICAgICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAibmFtZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN1YmplY3RzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAgICAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInJldGVudGlvbiI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAibGltaXRzIiwKICAgICAgICAgICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAgICAgICAgICJ3b3JrcXVldWUiCiAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9hZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAic3RvcmFnZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgICAgICAgICAibWVtb3J5IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJmaWxlIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgICAgICAgICAibmV3IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm1heF9zdHJlYW1zIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbWF4aW11bSBudW1iZXIgb2Ygc3RyZWFtcyB0byBhbGxvdyB1c2luZyB0aGlzIFRlbXBsYXRlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IC0xCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9LAogICAgICAgICJzdHJlYW1zIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkxpc3Qgb2YgU3RyZWFtcyBtYW5hZ2VkIGJ5IHRoaXMgVGVtcGxhdGUiLAogICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgImVycm9yIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiZXJyb3IiOiB7CiAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29kZSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImNvZGUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDMwMCwKICAgICAgICAgICAgICAibWF4aW11bSI6IDY5OQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGVzY3JpcHRpb24iOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3RlbXBsYXRlX2NyZWF0ZV9yZXNwb25zZSIKICAgIH0KICB9Cn0K")
 	schemas["io.nats.jetstream.api.v1.stream_template_delete_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9kZWxldGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLlRFTVBMQVRFLkRFTEVURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3RlbXBsYXRlX2RlbGV0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgewogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgInN1Y2Nlc3MiCiAgICAgIF0sCiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJzdWNjZXNzIjogewogICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIKICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3RlbXBsYXRlX2RlbGV0ZV9yZXNwb25zZSIKICAgIH0KICB9Cn0K")
-	schemas["io.nats.jetstream.api.v1.stream_template_info_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9pbmZvX3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5URU1QTEFURS5JTkZPIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfaW5mb19yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdHJlYW1zIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiY29uZmlnIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAibmFtZSIsCiAgICAgICAgICAgICJjb25maWciLAogICAgICAgICAgICAibWF4X3N0cmVhbXMiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBUZW1wbGF0ZSIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRlbXBsYXRlIGNvbmZpZ3VyYXRpb24gdG8gY3JlYXRlIFN0cmVhbXMgd2l0aCIsCiAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAic3ViamVjdHMiLAogICAgICAgICAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyIsCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiLAogICAgICAgICAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgICAgICAgICAibWF4X2FnZSIsCiAgICAgICAgICAgICAgICAic3RvcmFnZSIsCiAgICAgICAgICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAibmFtZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN1YmplY3RzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAgICAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInJldGVudGlvbiI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAibGltaXRzIiwKICAgICAgICAgICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAgICAgICAgICJ3b3JrcXVldWUiCiAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9hZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAic3RvcmFnZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgICAgICAgICAibWVtb3J5IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJtZW1vcnkiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IHJlcGxpY2FzIHRvIGtlZXAgZm9yIGVhY2ggbWVzc2FnZS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAxLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IDEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibm9fYWNrIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGlzYWJsZXMgYWNrbm93bGVkZ2luZyBtZXNzYWdlcyB0aGF0IGFyZSByZWNlaXZlZCBieSB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImJvb2xlYW4iLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IGZhbHNlCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInRlbXBsYXRlX293bmVyIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiB0aGUgU3RyZWFtIGlzIG1hbmFnZWQgYnkgYSBTdHJlYW0gVGVtcGxhdGUgdGhpcyBpZGVudGlmaWVzIHRoZSB0ZW1wbGF0ZSB0aGF0IG1hbmFnZXMgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImRpc2NhcmQiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIGEgU3RyZWFtIHJlYWNoIGl0J3MgbGltaXRzIGVpdGhlciBvbGQgbWVzc2FnZXMgYXJlIGRlbGV0ZWQgb3IgbmV3IG9uZXMgYXJlIGRlbmllZCIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICJvbGQiLAogICAgICAgICAgICAgICAgICAgICJuZXciCiAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogIm9sZCIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZHVwbGljYXRlX3dpbmRvdyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lIHdpbmRvdyB0byB0cmFjayBkdXBsaWNhdGUgbWVzc2FnZXMgZm9yLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciBkZWZhdWx0IiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgfQogICAgICAgICAgICB9LAogICAgICAgICAgICAibWF4X3N0cmVhbXMiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBtYXhpbXVtIG51bWJlciBvZiBzdHJlYW1zIHRvIGFsbG93IHVzaW5nIHRoaXMgVGVtcGxhdGUiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogLTEKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0sCiAgICAgICAgInN0cmVhbXMiOiB7CiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTGlzdCBvZiBTdHJlYW1zIG1hbmFnZWQgYnkgdGhpcyBUZW1wbGF0ZSIsCiAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiZXJyb3IiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJlcnJvciI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJjb2RlIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSFRUUCBsaWtlIGVycm9yIGNvZGUgaW4gdGhlIDMwMCB0byA1MDAgcmFuZ2UiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgICAgICJtYXhpbXVtIjogNjk5CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGh1bWFuIGZyaWVuZGx5IGRlc2NyaXB0aW9uIG9mIHRoZSBlcnJvciIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfaW5mb19yZXNwb25zZSIKICAgIH0KICB9Cn0K")
+	schemas["io.nats.jetstream.api.v1.stream_template_info_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9pbmZvX3Jlc3BvbnNlLmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlc3BvbnNlIGZyb20gdGhlIEpldFN0cmVhbSAkSlMuQVBJLlNUUkVBTS5URU1QTEFURS5JTkZPIEFQSSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfaW5mb19yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdHJlYW1zIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiY29uZmlnIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAibmFtZSIsCiAgICAgICAgICAgICJjb25maWciLAogICAgICAgICAgICAibWF4X3N0cmVhbXMiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJuYW1lIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBUZW1wbGF0ZSIKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRlbXBsYXRlIGNvbmZpZ3VyYXRpb24gdG8gY3JlYXRlIFN0cmVhbXMgd2l0aCIsCiAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAic3ViamVjdHMiLAogICAgICAgICAgICAgICAgInJldGVudGlvbiIsCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyIsCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiLAogICAgICAgICAgICAgICAgIm1heF9ieXRlcyIsCiAgICAgICAgICAgICAgICAibWF4X2FnZSIsCiAgICAgICAgICAgICAgICAic3RvcmFnZSIsCiAgICAgICAgICAgICAgICAibnVtX3JlcGxpY2FzIgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAibmFtZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgdW5pcXVlIG5hbWUgZm9yIHRoZSBTdHJlYW0sIGVtcHR5IGZvciBTdHJlYW0gVGVtcGxhdGVzLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJwYXR0ZXJuIjogIl5bXi4qPl0qJCIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN1YmplY3RzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBsaXN0IG9mIHN1YmplY3RzIHRvIGNvbnN1bWUsIHN1cHBvcnRzIHdpbGRjYXJkcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJhcnJheSIsCiAgICAgICAgICAgICAgICAgICJtaW5MZW5ndGgiOiAxLAogICAgICAgICAgICAgICAgICAiaXRlbXMiOiB7CiAgICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMQogICAgICAgICAgICAgICAgICB9CiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInJldGVudGlvbiI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtZXNzYWdlcyBhcmUgcmV0YWluZWQgaW4gdGhlIFN0cmVhbSwgb25jZSB0aGlzIGlzIGV4Y2VlZGVkIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAibGltaXRzIiwKICAgICAgICAgICAgICAgICAgICAiaW50ZXJlc3QiLAogICAgICAgICAgICAgICAgICAgICJ3b3JrcXVldWUiCiAgICAgICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogImxpbWl0cyIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2NvbnN1bWVycyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhvdyBtYW55IENvbnN1bWVycyBjYW4gYmUgZGVmaW5lZCBmb3IgYSBnaXZlbiBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X21zZ3MiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBtZXNzYWdlcyBtYXkgYmUgaW4gYSBTdHJlYW0sIG9sZGVzdCBtZXNzYWdlcyB3aWxsIGJlIHJlbW92ZWQgaWYgdGhlIFN0cmVhbSBleGNlZWRzIHRoaXMgc2l6ZS4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgYmlnIHRoZSBTdHJlYW0gbWF5IGJlLCB3aGVuIHRoZSBjb21iaW5lZCBzdHJlYW0gc2l6ZSBleGNlZWRzIHRoaXMgb2xkIG1lc3NhZ2VzIGFyZSByZW1vdmVkLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9hZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJNYXhpbXVtIGFnZSBvZiBhbnkgbWVzc2FnZSBpbiB0aGUgc3RyZWFtLCBleHByZXNzZWQgaW4gbWljcm9zZWNvbmRzLiAwIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMCwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dfc2l6ZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBsYXJnZXN0IG1lc3NhZ2UgdGhhdCB3aWxsIGJlIGFjY2VwdGVkIGJ5IHRoZSBTdHJlYW0uIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAic3RvcmFnZSI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBzdG9yYWdlIGJhY2tlbmQgdG8gdXNlIGZvciB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJlbnVtIjogWwogICAgICAgICAgICAgICAgICAgICJmaWxlIiwKICAgICAgICAgICAgICAgICAgICAibWVtb3J5IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJmaWxlIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgICAgICAgICAibmV3IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfSwKICAgICAgICAgICAgIm1heF9zdHJlYW1zIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbWF4aW11bSBudW1iZXIgb2Ygc3RyZWFtcyB0byBhbGxvdyB1c2luZyB0aGlzIFRlbXBsYXRlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IC0xCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9LAogICAgICAgICJzdHJlYW1zIjogewogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkxpc3Qgb2YgU3RyZWFtcyBtYW5hZ2VkIGJ5IHRoaXMgVGVtcGxhdGUiLAogICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciCiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgImVycm9yIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiZXJyb3IiOiB7CiAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29kZSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImNvZGUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDMwMCwKICAgICAgICAgICAgICAibWF4aW11bSI6IDY5OQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGVzY3JpcHRpb24iOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdLAogICJwcm9wZXJ0aWVzIjogewogICAgInR5cGUiOiB7CiAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICJjb25zdCI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3RlbXBsYXRlX2luZm9fcmVzcG9uc2UiCiAgICB9CiAgfQp9Cg==")
 	schemas["io.nats.jetstream.api.v1.stream_template_names_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9uYW1lc19yZXNwb25zZS5qc29uIiwKICAiZGVzY3JpcHRpb24iOiAiQSByZXNwb25zZSBmcm9tIHRoZSBKZXRTdHJlYW0gJEpTLkFQSS5TVFJFQU0uVEVNUExBVEUuTkFNRVMgQVBJIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV90ZW1wbGF0ZV9uYW1lc19yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJhbGxPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJ0b3RhbCIsCiAgICAgICAgIm9mZnNldCIsCiAgICAgICAgImxpbWl0IgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAidG90YWwiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0sCiAgICAgICAgIm9mZnNldCI6IHsKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgfSwKICAgICAgICAibGltaXQiOiB7CiAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgIm9uZU9mIjogWwogICAgewogICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgImVycm9yIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiZXJyb3IiOiB7CiAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAiY29kZSIKICAgICAgICAgIF0sCiAgICAgICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAgICAgImNvZGUiOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICAgICAibWluaW11bSI6IDMwMCwKICAgICAgICAgICAgICAibWF4aW11bSI6IDY5OQogICAgICAgICAgICB9LAogICAgICAgICAgICAiZGVzY3JpcHRpb24iOiB7CiAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgICAgIH0KICAgICAgICAgIH0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAic3RyZWFtcyIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbnN1bWVycyI6IHsKICAgICAgICAgICJ0eXBlIjogImFycmF5IiwKICAgICAgICAgICJpdGVtcyI6IHsKICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdGVtcGxhdGVfbmFtZXNfcmVzcG9uc2UiCiAgICB9CiAgfQp9Cg==")
 	schemas["io.nats.jetstream.api.v1.stream_template_names_request"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV90ZW1wbGF0ZV9uYW1lc19yZXF1ZXN0Lmpzb24iLAogICJkZXNjcmlwdGlvbiI6ICJBIHJlcXVlc3QgdG8gdGhlIEpldFN0cmVhbSAkSlMuQVBJLkNPTlNVTUVSLkxJU1QgQVBJIiwKICAidGl0bGUiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV90ZW1wbGF0ZV9uYW1lc19yZXF1ZXN0IiwKICAidHlwZSI6ICJvYmplY3QiLAogICJhbGxPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgImFkZGl0aW9uYWxQcm9wZXJ0aWVzIjogZmFsc2UsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAib2Zmc2V0IgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAib2Zmc2V0IjogewogICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICB9CiAgICAgIH0KICAgIH0KICBdCn0K")
-	schemas["io.nats.jetstream.api.v1.stream_update_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV91cGRhdGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLlVQREFURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3VwZGF0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdGF0ZSIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICJhbGxPZiI6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAibWVtb3J5IgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJudW1fcmVwbGljYXMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSByZXBsaWNhcyB0byBrZWVwIGZvciBlYWNoIG1lc3NhZ2UuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAxCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm5vX2FjayI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkRpc2FibGVzIGFja25vd2xlZGdpbmcgbWVzc2FnZXMgdGhhdCBhcmUgcmVjZWl2ZWQgYnkgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJib29sZWFuIiwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJ0ZW1wbGF0ZV9vd25lciI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gdGhlIFN0cmVhbSBpcyBtYW5hZ2VkIGJ5IGEgU3RyZWFtIFRlbXBsYXRlIHRoaXMgaWRlbnRpZmllcyB0aGUgdGVtcGxhdGUgdGhhdCBtYW5hZ2VzIHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkaXNjYXJkIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiV2hlbiBhIFN0cmVhbSByZWFjaCBpdCdzIGxpbWl0cyBlaXRoZXIgb2xkIG1lc3NhZ2VzIGFyZSBkZWxldGVkIG9yIG5ldyBvbmVzIGFyZSBkZW5pZWQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAib2xkIiwKICAgICAgICAgICAgICAgICAgICAibmV3IgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJvbGQiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImR1cGxpY2F0ZV93aW5kb3ciOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZSB3aW5kb3cgdG8gdHJhY2sgZHVwbGljYXRlIG1lc3NhZ2VzIGZvciwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgZGVmYXVsdCIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfQogICAgICAgICAgXQogICAgICAgIH0sCiAgICAgICAgInN0YXRlIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEZXRhaWwgYWJvdXQgdGhlIGN1cnJlbnQgU3RhdGUgb2YgdGhlIFN0cmVhbSIsCiAgICAgICAgICAiYWxsT2YiOiBbCiAgICAgICAgICAgIHsKICAgICAgICAgICAgICAidHlwZSI6ICJvYmplY3QiLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgICAgICJtZXNzYWdlcyIsCiAgICAgICAgICAgICAgICAiYnl0ZXMiLAogICAgICAgICAgICAgICAgImZpcnN0X3NlcSIsCiAgICAgICAgICAgICAgICAibGFzdF9zZXEiLAogICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IgogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICAgICAibWVzc2FnZXMiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTnVtYmVyIG9mIG1lc3NhZ2VzIHN0b3JlZCBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImJ5dGVzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkNvbWJpbmVkIHNpemUgb2YgYWxsIG1lc3NhZ2VzIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZmlyc3Rfc2VxIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlNlcXVlbmNlIG51bWJlciBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImZpcnN0X3RzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWVzdGFtcCBvZiB0aGUgZmlyc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImxhc3Rfc2VxIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlNlcXVlbmNlIG51bWJlciBvZiB0aGUgbGFzdCBtZXNzYWdlIGluIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibGFzdF90cyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lc3RhbXAgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgImZvcm1hdCI6ICJkYXRlLXRpbWUiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImNvbnN1bWVyX2NvdW50IjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBDb25zdW1lcnMgYXR0YWNoZWQgdG8gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIH0KICAgICAgICAgICAgfQogICAgICAgICAgXQogICAgICAgIH0KICAgICAgfQogICAgfSwKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJlcnJvciIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImVycm9yIjogewogICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAgICAgImNvZGUiCiAgICAgICAgICBdLAogICAgICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgICAgICJjb2RlIjogewogICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIVFRQIGxpa2UgZXJyb3IgY29kZSBpbiB0aGUgMzAwIHRvIDUwMCByYW5nZSIsCiAgICAgICAgICAgICAgIm1pbmltdW0iOiAzMDAsCiAgICAgICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICAgICAgfSwKICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogewogICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgaHVtYW4gZnJpZW5kbHkgZGVzY3JpcHRpb24gb2YgdGhlIGVycm9yIgogICAgICAgICAgICB9CiAgICAgICAgICB9CiAgICAgICAgfQogICAgICB9CiAgICB9CiAgXSwKICAicHJvcGVydGllcyI6IHsKICAgICJ0eXBlIjogewogICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAiY29uc3QiOiAiaW8ubmF0cy5qZXRzdHJlYW0uYXBpLnYxLnN0cmVhbV91cGRhdGVfcmVzcG9uc2UiCiAgICB9CiAgfQp9Cg==")
+	schemas["io.nats.jetstream.api.v1.stream_update_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3N0cmVhbV91cGRhdGVfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgZnJvbSB0aGUgSmV0U3RyZWFtICRKUy5BUEkuU1RSRUFNLlVQREFURSBBUEkiLAogICJ0aXRsZSI6ICJpby5uYXRzLmpldHN0cmVhbS5hcGkudjEuc3RyZWFtX3VwZGF0ZV9yZXNwb25zZSIsCiAgInR5cGUiOiAib2JqZWN0IiwKICAicmVxdWlyZWQiOiBbCiAgICAidHlwZSIKICBdLAogICJvbmVPZiI6IFsKICAgIHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb25maWciLAogICAgICAgICJzdGF0ZSIKICAgICAgXSwKICAgICAgInByb3BlcnRpZXMiOiB7CiAgICAgICAgImNvbmZpZyI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIGFjdGl2ZSBjb25maWd1cmF0aW9uIGZvciB0aGUgU3RyZWFtIiwKICAgICAgICAgICJhbGxPZiI6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICAgICAgICAgInN1YmplY3RzIiwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iLAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiLAogICAgICAgICAgICAgICAgIm1heF9tc2dzIiwKICAgICAgICAgICAgICAgICJtYXhfYnl0ZXMiLAogICAgICAgICAgICAgICAgIm1heF9hZ2UiLAogICAgICAgICAgICAgICAgInN0b3JhZ2UiLAogICAgICAgICAgICAgICAgIm51bV9yZXBsaWNhcyIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJhZGRpdGlvbmFsUHJvcGVydGllcyI6IGZhbHNlLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm5hbWUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIHVuaXF1ZSBuYW1lIGZvciB0aGUgU3RyZWFtLCBlbXB0eSBmb3IgU3RyZWFtIFRlbXBsYXRlcy4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAicGF0dGVybiI6ICJeW14uKj5dKiQiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJzdWJqZWN0cyI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIkEgbGlzdCBvZiBzdWJqZWN0cyB0byBjb25zdW1lLCBzdXBwb3J0cyB3aWxkY2FyZHMuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYXJyYXkiLAogICAgICAgICAgICAgICAgICAibWluTGVuZ3RoIjogMSwKICAgICAgICAgICAgICAgICAgIml0ZW1zIjogewogICAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICAgIm1pbkxlbmd0aCI6IDEKICAgICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJyZXRlbnRpb24iOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWVzc2FnZXMgYXJlIHJldGFpbmVkIGluIHRoZSBTdHJlYW0sIG9uY2UgdGhpcyBpcyBleGNlZWRlZCBvbGQgbWVzc2FnZXMgYXJlIHJlbW92ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgImxpbWl0cyIsCiAgICAgICAgICAgICAgICAgICAgImludGVyZXN0IiwKICAgICAgICAgICAgICAgICAgICAid29ya3F1ZXVlIgogICAgICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6ICJsaW1pdHMiCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9jb25zdW1lcnMiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJIb3cgbWFueSBDb25zdW1lcnMgY2FuIGJlIGRlZmluZWQgZm9yIGEgZ2l2ZW4gU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgIm1heF9tc2dzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgbWVzc2FnZXMgbWF5IGJlIGluIGEgU3RyZWFtLCBvbGRlc3QgbWVzc2FnZXMgd2lsbCBiZSByZW1vdmVkIGlmIHRoZSBTdHJlYW0gZXhjZWVkcyB0aGlzIHNpemUuIC0xIGZvciB1bmxpbWl0ZWQuIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogLTEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogLTEKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibWF4X2J5dGVzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IGJpZyB0aGUgU3RyZWFtIG1heSBiZSwgd2hlbiB0aGUgY29tYmluZWQgc3RyZWFtIHNpemUgZXhjZWVkcyB0aGlzIG9sZCBtZXNzYWdlcyBhcmUgcmVtb3ZlZC4gLTEgZm9yIHVubGltaXRlZC4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAtMSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAtMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfYWdlIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiTWF4aW11bSBhZ2Ugb2YgYW55IG1lc3NhZ2UgaW4gdGhlIHN0cmVhbSwgZXhwcmVzc2VkIGluIG1pY3Jvc2Vjb25kcy4gMCBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJtYXhfbXNnX3NpemUiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgbGFyZ2VzdCBtZXNzYWdlIHRoYXQgd2lsbCBiZSBhY2NlcHRlZCBieSB0aGUgU3RyZWFtLiAtMSBmb3IgdW5saW1pdGVkLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IC0xLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IC0xCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgInN0b3JhZ2UiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgc3RvcmFnZSBiYWNrZW5kIHRvIHVzZSBmb3IgdGhlIFN0cmVhbS4iLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICAgICAiZW51bSI6IFsKICAgICAgICAgICAgICAgICAgICAiZmlsZSIsCiAgICAgICAgICAgICAgICAgICAgIm1lbW9yeSIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAiZmlsZSIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAibnVtX3JlcGxpY2FzIjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSG93IG1hbnkgcmVwbGljYXMgdG8ga2VlcCBmb3IgZWFjaCBtZXNzYWdlLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDEsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogMQogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJub19hY2siOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJEaXNhYmxlcyBhY2tub3dsZWRnaW5nIG1lc3NhZ2VzIHRoYXQgYXJlIHJlY2VpdmVkIGJ5IHRoZSBTdHJlYW0uIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICAgICAgICAgICAgICJkZWZhdWx0IjogZmFsc2UKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAidGVtcGxhdGVfb3duZXIiOiB7CiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJXaGVuIHRoZSBTdHJlYW0gaXMgbWFuYWdlZCBieSBhIFN0cmVhbSBUZW1wbGF0ZSB0aGlzIGlkZW50aWZpZXMgdGhlIHRlbXBsYXRlIHRoYXQgbWFuYWdlcyB0aGUgU3RyZWFtLiIsCiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAiZGlzY2FyZCI6IHsKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIldoZW4gYSBTdHJlYW0gcmVhY2ggaXQncyBsaW1pdHMgZWl0aGVyIG9sZCBtZXNzYWdlcyBhcmUgZGVsZXRlZCBvciBuZXcgb25lcyBhcmUgZGVuaWVkIiwKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImVudW0iOiBbCiAgICAgICAgICAgICAgICAgICAgIm9sZCIsCiAgICAgICAgICAgICAgICAgICAgIm5ldyIKICAgICAgICAgICAgICAgICAgXSwKICAgICAgICAgICAgICAgICAgImRlZmF1bHQiOiAib2xkIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJkdXBsaWNhdGVfd2luZG93IjogewogICAgICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiVGhlIHRpbWUgd2luZG93IHRvIHRyYWNrIGR1cGxpY2F0ZSBtZXNzYWdlcyBmb3IsIGV4cHJlc3NlZCBpbiBtaWNyb3NlY29uZHMuIDAgZm9yIGRlZmF1bHQiLAogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwLAogICAgICAgICAgICAgICAgICAiZGVmYXVsdCI6IDAKICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0KICAgICAgICAgIF0KICAgICAgICB9LAogICAgICAgICJzdGF0ZSI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiRGV0YWlsIGFib3V0IHRoZSBjdXJyZW50IFN0YXRlIG9mIHRoZSBTdHJlYW0iLAogICAgICAgICAgImFsbE9mIjogWwogICAgICAgICAgICB7CiAgICAgICAgICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgICAgICAgICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAgICAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICAgICAibWVzc2FnZXMiLAogICAgICAgICAgICAgICAgImJ5dGVzIiwKICAgICAgICAgICAgICAgICJmaXJzdF9zZXEiLAogICAgICAgICAgICAgICAgImxhc3Rfc2VxIiwKICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCIKICAgICAgICAgICAgICBdLAogICAgICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAgICAgIm1lc3NhZ2VzIjogewogICAgICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIk51bWJlciBvZiBtZXNzYWdlcyBzdG9yZWQgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJieXRlcyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJDb21iaW5lZCBzaXplIG9mIGFsbCBtZXNzYWdlcyBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImZpcnN0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGZpcnN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJtaW5pbXVtIjogMAogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJmaXJzdF90cyI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgICAgICAgICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSB0aW1lc3RhbXAgb2YgdGhlIGZpcnN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJsYXN0X3NlcSI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJTZXF1ZW5jZSBudW1iZXIgb2YgdGhlIGxhc3QgbWVzc2FnZSBpbiB0aGUgU3RyZWFtIiwKICAgICAgICAgICAgICAgICAgIm1pbmltdW0iOiAwCiAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgImxhc3RfdHMiOiB7CiAgICAgICAgICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJUaGUgdGltZXN0YW1wIG9mIHRoZSBsYXN0IG1lc3NhZ2UgaW4gdGhlIFN0cmVhbSIsCiAgICAgICAgICAgICAgICAgICJmb3JtYXQiOiAiZGF0ZS10aW1lIgogICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICJjb25zdW1lcl9jb3VudCI6IHsKICAgICAgICAgICAgICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJOdW1iZXIgb2YgQ29uc3VtZXJzIGF0dGFjaGVkIHRvIHRoZSBTdHJlYW0iLAogICAgICAgICAgICAgICAgICAibWluaW11bSI6IDAKICAgICAgICAgICAgICAgIH0KICAgICAgICAgICAgICB9CiAgICAgICAgICAgIH0KICAgICAgICAgIF0KICAgICAgICB9CiAgICAgIH0KICAgIH0sCiAgICB7CiAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICJyZXF1aXJlZCI6IFsKICAgICAgICAiZXJyb3IiCiAgICAgIF0sCiAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICJlcnJvciI6IHsKICAgICAgICAgICJ0eXBlIjogIm9iamVjdCIsCiAgICAgICAgICAicmVxdWlyZWQiOiBbCiAgICAgICAgICAgICJjb2RlIgogICAgICAgICAgXSwKICAgICAgICAgICJwcm9wZXJ0aWVzIjogewogICAgICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJpbnRlZ2VyIiwKICAgICAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiSFRUUCBsaWtlIGVycm9yIGNvZGUgaW4gdGhlIDMwMCB0byA1MDAgcmFuZ2UiLAogICAgICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgICAgICJtYXhpbXVtIjogNjk5CiAgICAgICAgICAgIH0sCiAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICAgICAidHlwZSI6ICJzdHJpbmciLAogICAgICAgICAgICAgICJkZXNjcmlwdGlvbiI6ICJBIGh1bWFuIGZyaWVuZGx5IGRlc2NyaXB0aW9uIG9mIHRoZSBlcnJvciIKICAgICAgICAgICAgfQogICAgICAgICAgfQogICAgICAgIH0KICAgICAgfQogICAgfQogIF0sCiAgInByb3BlcnRpZXMiOiB7CiAgICAidHlwZSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImNvbnN0IjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5zdHJlYW1fdXBkYXRlX3Jlc3BvbnNlIgogICAgfQogIH0KfQo=")
 	schemas["io.nats.jetstream.api.v1.pub_ack_response"], _ = base64.StdEncoding.DecodeString("ewogICIkc2NoZW1hIjogImh0dHA6Ly9qc29uLXNjaGVtYS5vcmcvZHJhZnQtMDcvc2NoZW1hIyIsCiAgIiRpZCI6ICJodHRwczovL25hdHMuaW8vc2NoZW1hcy9qZXRzdHJlYW0vYXBpL3YxL3B1Yl9hY2tfcmVzcG9uc2UuanNvbiIsCiAgImRlc2NyaXB0aW9uIjogIkEgcmVzcG9uc2UgcmVjZWl2ZWQgd2hlbiBwdWJsaXNoaW5nIGEgbWVzc2FnZSIsCiAgInRpdGxlIjogImlvLm5hdHMuamV0c3RyZWFtLmFwaS52MS5wdWJfYWNrX3Jlc3BvbnNlIiwKICAidHlwZSI6ICJvYmplY3QiLAogICJyZXF1aXJlZCI6IFsKICAgICJzdHJlYW0iCiAgXSwKICAiYWRkaXRpb25hbFByb3BlcnRpZXMiOiBmYWxzZSwKICAicHJvcGVydGllcyI6IHsKICAgICJlcnJvciI6IHsKICAgICAgInR5cGUiOiAib2JqZWN0IiwKICAgICAgInJlcXVpcmVkIjogWwogICAgICAgICJjb2RlIgogICAgICBdLAogICAgICAicHJvcGVydGllcyI6IHsKICAgICAgICAiY29kZSI6IHsKICAgICAgICAgICJ0eXBlIjogImludGVnZXIiLAogICAgICAgICAgImRlc2NyaXB0aW9uIjogIkhUVFAgbGlrZSBlcnJvciBjb2RlIGluIHRoZSAzMDAgdG8gNTAwIHJhbmdlIiwKICAgICAgICAgICJtaW5pbXVtIjogMzAwLAogICAgICAgICAgIm1heGltdW0iOiA2OTkKICAgICAgICB9LAogICAgICAgICJkZXNjcmlwdGlvbiI6IHsKICAgICAgICAgICJ0eXBlIjogInN0cmluZyIsCiAgICAgICAgICAiZGVzY3JpcHRpb24iOiAiQSBodW1hbiBmcmllbmRseSBkZXNjcmlwdGlvbiBvZiB0aGUgZXJyb3IiCiAgICAgICAgfQogICAgICB9CiAgICB9LAogICAgInN0cmVhbSI6IHsKICAgICAgInR5cGUiOiAic3RyaW5nIiwKICAgICAgImRlc2NyaXB0aW9uIjogIlRoZSBuYW1lIG9mIHRoZSBzdHJlYW0gdGhhdCByZWNlaXZlZCB0aGUgbWVzc2FnZSIsCiAgICAgICJtaW5MZW5ndGgiOiAxCiAgICB9LAogICAgInNlcSI6IHsKICAgICAgInR5cGUiOiAiaW50ZWdlciIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJJZiBzdWNjZXNzZnVsIHRoaXMgd2lsbCBiZSB0aGUgc2VxdWVuY2UgdGhlIG1lc3NhZ2UgaXMgc3RvcmVkIGF0IiwKICAgICAgIm1pbmltdW0iOiAwLAogICAgICAibWF4aW11bSI6IDE4NDQ2NzQ0MDczNzA5NTUyMDAwCiAgICB9LAogICAgImR1cGxpY2F0ZSI6IHsKICAgICAgInR5cGUiOiAiYm9vbGVhbiIsCiAgICAgICJkZXNjcmlwdGlvbiI6ICJJbmRpY2F0ZXMgdGhhdCB0aGUgbWVzc2FnZSB3YXMgbm90IHN0b3JlZCBkdWUgdG8gdGhlIE5hdHMtTXNnLUlkIGhlYWRlciBhbmQgZHVwbGljYXRlIHRyYWNraW5nIiwKICAgICAgImRlZmF1bHQiOiBmYWxzZQogICAgfQogIH0KfQo=")
 }
